% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agfd-package.R
\docType{package}
\name{abares-package}
\alias{abares}
\alias{abares-package}
\title{abares: Provides automated downloading, parsing and importing of Australian Bureau of Agricultural and Resource Economics and Sciences data sources.}
\description{
Fetch, import and query data from the Australian Bureau of Agricultural and Resource Economics and Sciences including the Historical National Estimates, Historical State Estimates, Historical Regional Estimates, Estimates by Size, Estimates by Performance Category, \url{https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download}; Australian Gridded Farm Data set, \url{https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data}; and Australian Agricultural and Grazing Industries Survey (AAGIS) region mapping files.
}
\keyword{internal}
