% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_winter_broadacre_crops.R
\name{abares_crop_data}
\alias{abares_crop_data}
\title{Create a unified data.table object of all sheets in Excel Workbook}
\usage{
abares_crop_data(filename)
}
\arguments{
\item{filename}{The name of the Excel workbook for import}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} object.
}
\description{
Create a unified data.table object of all sheets in Excel Workbook
}
\keyword{internal}
