% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agfd-package.R
\docType{package}
\name{agfd-package}
\alias{agfd}
\alias{agfd-package}
\title{agfd: Provides automated downloading, parsing and importing of Australian Gridded Farm Data from Australian Department of Agriculture, Fisheries and Forestry.}
\description{
Fetch, import and query data from the Australian Gridded Farm Data set, \url{https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data}
}
\keyword{internal}
