% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{build_httr2_request}
\alias{build_httr2_request}
\title{Build a base httr2 request with polite defaults (internal)}
\usage{
build_httr2_request(url)
}
\arguments{
\item{url}{Character scalar; the request URL.}
}
\value{
An \code{httr2_request} object.
}
\description{
Build a base httr2 request with polite defaults (internal)
}
\keyword{internal}
