% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_thickness.R
\name{display_soil_thickness_metadata}
\alias{display_soil_thickness_metadata}
\title{View Complete Metadata Associated with Soil Thickness Data}
\usage{
display_soil_thickness_metadata(x)
}
\arguments{
\item{x}{An \code{abares.soil.thickness.files} object}
}
\value{
Nothing, called for its side effects, it prints the complete
metadata file to the \R console
}
\description{
View Complete Metadata Associated with Soil Thickness Data
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
display_soil_thickness_metadata(get_soil_thickness(cache = TRUE))
\dontshow{\}) # examplesIf}
}
