% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.apply_conditional_options}
\alias{.apply_conditional_options}
\title{Apply conditional options to httr2 request}
\usage{
.apply_conditional_options(req, url)
}
\arguments{
\item{req}{An httr2 request object}

\item{url}{The URL being requested}
}
\value{
Modified httr2 request object
}
\description{
Apply conditional options to httr2 request
}
\keyword{internal}
