% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.apply_timeouts}
\alias{.apply_timeouts}
\title{Apply timeouts to httr2 request.}
\usage{
.apply_timeouts(req, timeouts)
}
\arguments{
\item{req}{httr2_request object.}

\item{timeouts}{List of timeout values from \code{.get_timeouts()}.}
}
\value{
Modified httr2_request object.
}
\description{
Apply timeouts to httr2 request.
}
\keyword{internal}
