% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.build_request}
\alias{.build_request}
\title{Build httr2 request with standard headers}
\usage{
.build_request(url)
}
\arguments{
\item{url}{Request URL.}
}
\value{
httr2_request object.
}
\description{
Build httr2 request with standard headers
}
\keyword{internal}
