% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.bump_timeouts_by_context}
\alias{.bump_timeouts_by_context}
\title{Bump timeouts by host/type context (internal)}
\usage{
.bump_timeouts_by_context(t_o, url, probe)
}
\arguments{
\item{t_o}{timeouts list from \code{.dataset_timeouts()}}

\item{url}{original URL}

\item{probe}{list from \code{.head_or_range_probe()}}
}
\value{
possibly adjusted list t_o
}
\description{
Bump timeouts by host/type context (internal)
}
\keyword{internal}
