% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agfd_dt.R
\name{.check_agfd_yyyy}
\alias{.check_agfd_yyyy}
\title{Check AGFD years for validity}
\usage{
.check_agfd_yyyy(.yyyy)
}
\arguments{
\item{yyyy}{A year value for checking}
}
\description{
Check AGFD years for validity
}
\keyword{internal}
