% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.check_class}
\alias{.check_class}
\title{Check for read.abares.something S3 class}
\usage{
.check_class(x, class)
}
\arguments{
\item{x}{An object for validating.}

\item{class}{An S3 class to validate against.}
}
\value{
An invisible \code{NULL}, called for its side-effects of class
validation.
}
\description{
Check for read.abares.something S3 class
}
\keyword{internal}
