% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.check_clum_inputs}
\alias{.check_clum_inputs}
\title{Check CLUM inputs for stars and terra Importing}
\usage{
.check_clum_inputs(data_set, x)
}
\arguments{
\item{data_set}{A user provided string value.}

\item{x}{A user provided string value.}
}
\value{
A \code{list} object containing a validated \code{data_set} value and \code{x}
object (a geotiff object file path for importing).
}
\description{
Check CLUM inputs for stars and terra Importing
}
\keyword{internal}
