% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aagis_regions.R
\name{.check_existing_aagis}
\alias{.check_existing_aagis}
\title{Check for a pre-existing file before downloading}
\usage{
.check_existing_aagis(cache)
}
\value{
An \CRANpkg{sf} object of AAGIS regions.
}
\description{
Checks the user cache first, then \code{tempdir()} for the files before
returning a \code{NULL} value. If \code{cache == TRUE} and the file is not in the user
cache, but is in \code{tempdir()}, it is saved to the cache before being returned
in the current session.
}
\keyword{internal}
