% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topsoil_thickness.R
\name{.convert_and_copy_files}
\alias{.convert_and_copy_files}
\title{Convert files to GTiff and copy metadata}
\usage{
.convert_and_copy_files(.files_path)
}
\arguments{
\item{.files_path}{Filepath where topsoil files have been stored locally.}
}
\value{
A list of files that have been converted to GeoTIFF and stripped of
the raster attribute table and metadata for the GeoTIFF file.
}
\description{
Convert files to GTiff and copy metadata
}
\keyword{internal}
