% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agfd.R
\name{.copy_local_agfd_zip}
\alias{.copy_local_agfd_zip}
\title{Copies local AGFD file to tempdir() so the org file isn't touched}
\usage{
.copy_local_agfd_zip(x)
}
\arguments{
\item{x}{A filepath pointing to the AGFD file provided by the user}
}
\value{
A character vector of file paths for AGFD netCDF files
}
\description{
Copies local AGFD file to tempdir() so the org file isn't touched
}
\keyword{internal}
