% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clum.R
\name{.copy_local_clum_zip}
\alias{.copy_local_clum_zip}
\title{Copies local CLUM file to tempdir() so the org file isn't touched}
\usage{
.copy_local_clum_zip(x)
}
\arguments{
\item{x}{A filepath pointing to the CLUM file provided by the user}
}
\value{
A character vector of geotif file names.
}
\description{
Copies local CLUM file to tempdir() so the org file isn't touched
}
\keyword{internal}
