% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clum_terra.R
\name{.create_clum_50m_coltab}
\alias{.create_clum_50m_coltab}
\title{Create and Apply a Colour data.frame for the clum_50m_2023_v2 Data}
\usage{
.create_clum_50m_coltab()
}
\description{
Creates a \code{data.frame()} that contains the hexadecimal colour codes that
correspond with the integer values to display the map colours as published
by \acronym{ABARES} for the Catchment Level Land Use (\acronym{clum}) data.
Values are derived from the .qml file provided by \acronym{ABARES}.
}
\examples{
.apply_clum_50m_col_df()

}
\keyword{internal}
