% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_thickness.R
\name{.create_soil_thickness_list}
\alias{.create_soil_thickness_list}
\title{Create a object of read.abares.soil.thickness.files}
\usage{
.create_soil_thickness_list(soil_dir)
}
\arguments{
\item{dir}{File where files have been downloaded.}
}
\value{
A \code{read.abares.soil.thickness} object, which is a named \code{list} with
the fs::path_file of the resulting Esri Grid file and text file of metadata.
}
\description{
Create a object of read.abares.soil.thickness.files
}
\keyword{internal}
