% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topsoil_thickness.R
\name{.create_topsoil_thickness_list}
\alias{.create_topsoil_thickness_list}
\title{Create a object of read.abares.topsoil.thickness.files}
\usage{
.create_topsoil_thickness_list(topsoil_thickness_cache)
}
\arguments{
\item{dir}{File where files have been stored.}
}
\value{
A \code{read.abares.soil.thickness} object, which is a named \code{list} with
the \code{\link[fs:path_file]{fs::path_file()}} of the resulting GTiff file and text file of metadata.
}
\description{
Create a object of read.abares.topsoil.thickness.files
}
\keyword{internal}
