% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aagis_regions.R
\name{.download_aagis_shp}
\alias{.download_aagis_shp}
\title{Download the 'Australian Agricultural and Grazing Industries Survey' (AAGIS) regions shapefile}
\usage{
.download_aagis_shp(cache)
}
\arguments{
\item{cache}{\code{Boolean} Cache the \acronym{AAGIS} regions shape files after
download using \code{tools::R_user_dir("read.abares")} to identify the proper
directory for storing user data in a cache for this package. Defaults to
\code{TRUE}, caching the files locally as a native \R object. If \code{FALSE}, this
function uses \code{tempdir()} and the files are deleted upon closing of the \R
session.}
}
\value{
An \cranpkg{sf} object of AAGIS regions.
}
\description{
Handles downloading, caching (if requested) and importing of AAGIS regions
geospatial data.  The geometries are corrected for validity before returning
to the user.
}
\keyword{internal}
