% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abares_trade.R
\name{.download_abares_trade}
\alias{.download_abares_trade}
\title{Download the ABARES trade CSV file}
\usage{
.download_abares_trade(.cache)
}
\arguments{
\item{.cache}{\code{Boolean} enable caching?}
}
\value{
A \CRANpkg{data.table} object of the \acronym{ABARES} trade data.
}
\description{
Handles downloading and caching (if requested) of ABARES Trade data files.
}
\keyword{internal}
