% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.download_file}
\alias{.download_file}
\title{Wrap curl::curl_download mainly for ease of testing}
\usage{
.download_file(url, destfile, quiet)
}
\arguments{
\item{url}{Character the URL being requested.}

\item{destfile}{Character a filepath to be written to local storage.}

\item{quiet}{How verbose should the download be? Defaults to \code{TRUE}.}
}
\value{
Invisibly returns \code{NULL}, called for its side-effects, writes an
object to disk.
}
\description{
Wrap curl::curl_download mainly for ease of testing
}
\keyword{internal}
