% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topsoil_thickness.R
\name{.download_topsoil_thickness}
\alias{.download_topsoil_thickness}
\title{Downloads topsoil thickness data if not already found locally}
\usage{
.download_topsoil_thickness()
}
\value{
A list of the resulting data and text file of metadata after
downloading and upzipping.
}
\description{
Downloads topsoil thickness data if not already found locally
}
\keyword{internal}
