% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.ensure_timeout_floor}
\alias{.ensure_timeout_floor}
\title{Ensure a minimum total-timeout for large files (internal)}
\usage{
.ensure_timeout_floor(t_o, size_bytes)
}
\arguments{
\item{t_o}{timeouts list from \code{.dataset_timeouts()}}

\item{size_bytes}{numeric or NA}
}
\value{
possibly adjusted list t_o
}
\description{
Ensure a minimum total-timeout for large files (internal)
}
\keyword{internal}
