% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.extract_filename_from_cd}
\alias{.extract_filename_from_cd}
\title{Extract filename from Content-Disposition (internal)}
\usage{
.extract_filename_from_cd(content_disp)
}
\description{
Extract filename from Content-Disposition (internal)
}
\keyword{internal}
