% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_abs_production_data.R
\name{.find_years}
\alias{.find_years}
\title{Find Which Financial Years Data are Available}
\usage{
.find_years(data_set)
}
\value{
A string value of financial years that match availability from the
ABS website, \emph{e.g.}, \code{2023-24}.
}
\description{
Grabs the ABS website and uses a regexp to find what financial years are
available for download.
}
\keyword{internal}
