% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agfd.R
\name{.get_agfd}
\alias{.get_agfd}
\title{Get "Australian Gridded Farm Data" (AGFD)}
\usage{
.get_agfd(.fixed_prices, .yyyy)
}
\arguments{
\item{.fixed_prices}{Download historical climate and prices or historical
climate and fixed prices as described in (Hughes \emph{et al.} 2022).}

\item{.yyyy}{Returns only data for the specified year or years for climate
data (fixed prices) or the years for historical climate and prices depending
upon the setting of \code{.fixed_prices}.  Note that this will still download the
entire data set, that cannot be avoided, but will only return the
requested year(s) in your \R session.  Valid years are from 1991 to 2023
inclusive.}
}
\value{
A \code{list()} object, a list of NetCDF files containing the "Australian
Gridded Farm Data".
}
\description{
Used by the \code{read_agfd} family of functions, downloads the "Australian
Gridded Farm Data" (\acronym{AGFD}) data and unzips the compressed files to
NetCDF for importing.
}
\examples{
agfd <- get_agfd()

agfd

}
\keyword{internal}
