% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clum_data.R
\name{.get_clum}
\alias{.get_clum}
\title{Get Catchment Scale "Land Use of Australia" Data}
\source{
\url{https://10.25814/2w2p-ph98}.
}
\usage{
.get_clum(.data_set)
}
\arguments{
\item{.data_set}{A string value indicating the data desired for download.
One of:
\describe{
\item{clum_50m_2023_v2}{Catchment Scale Land Use of Australia – Update December 2023 version 2}
\item{scale_date_update}{Catchment Scale Land Use of Australia - Date and Scale of Mapping}
}.}
}
\value{
A list of files containing a spatial data file of or files related
to Australian catchment scale land use data.
}
\description{
An internal function used by \code{\link[=read_clum_terra]{read_clum_terra()}} and \code{\link[=read_clum_stars]{read_clum_stars()}} that
downloads catchment level land use data files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
CLUM50m <- get_clum(.data_set = "clum_50m_2023_v2")

CLUM50m
\dontshow{\}) # examplesIf}
}
\references{
ABARES 2024, Catchment Scale Land Use of Australia – Update December 2023
version 2, Australian Bureau of Agricultural and Resource Economics and
Sciences, Canberra, June, CC BY 4.0, DOI: \doi{10.25814/2w2p-ph98}.
}
\keyword{internal}
