% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lum_files.R
\name{.get_lum_files}
\alias{.get_lum_files}
\title{Unified CLUM and NLUM Checks and Downloading Automation}
\usage{
.get_lum_files(x, data_set, lum)
}
\arguments{
\item{data_set}{A user provided string value.}

\item{lum}{A string value indicating whether to process 'clum' or 'nlum'
data.}
}
\value{
A list object containing the path to the zip file and the filename
of the GeoTIFF contained in the zip file.
}
\description{
Checks user inputs for CLUM and NLUM functions and validates inputs. If
no file is provided, downloads the data set. If \code{x} is provided, checks that
the file exists and has the proper files in it. Calls the respective CLUM or
NLUM download function.
}
\examples{
# for NLUM data download
.get_lum_files(x = NULL, data_set = "Y202021", lum = "nlum")

# for CLUM data download
.get_lum_files(x = NULL, data_set = "clum_50m_2023_v2", lum = "clum")

}
\keyword{internal}
