% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.get_remote_file_size}
\alias{.get_remote_file_size}
\title{Check remote file size for comparison with cached-version}
\usage{
.get_remote_file_size(url)
}
\arguments{
\item{url}{A character \code{string} that provides the \acronym{url} for the file
that should be checked.}
}
\value{
A \code{numeric} value indicating the file size extracted from the
response headers.
}
\description{
Fetches remote file size before downloading the file.  Useful for checking
whether the file has changed since it was first cached.
}
\keyword{internal}
