% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.is_agriculture_url}
\alias{.is_agriculture_url}
\title{Check if URL is from agriculture.gov.au documents}
\usage{
.is_agriculture_url(url)
}
\arguments{
\item{url}{URL to check}
}
\value{
Logical
}
\description{
Check if URL is from agriculture.gov.au documents
}
\keyword{internal}
