% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topsoil_thickness.R
\name{.list_topsoil_thickness_files}
\alias{.list_topsoil_thickness_files}
\title{Create a list of topsoil thickness files for read_topsoil_thickness functions}
\usage{
.list_topsoil_thickness_files(.files_path)
}
\arguments{
\item{.files_path}{A character string pointed at the local storage where the
topsoil thickness files are found.}
}
\value{
A \code{list} of class \code{read.abares.topsoil.thickness.files}.
}
\description{
Create a list of topsoil thickness files for read_topsoil_thickness functions
}
\keyword{internal}
