% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.parse_url_info}
\alias{.parse_url_info}
\title{Parse URL information.}
\usage{
.parse_url_info(url)
}
\arguments{
\item{url}{Character string URL to parse.}
}
\value{
List with parsed URL components.
}
\description{
Parse URL information.
}
\keyword{internal}
