% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agfd.R
\name{.read_ncdf_from_zip}
\alias{.read_ncdf_from_zip}
\title{Unzip AGFD NetCDF files from ZIP}
\usage{
.read_ncdf_from_zip(zip_path)
}
\arguments{
\item{zip_path}{Path to the ZIP file containing NetCDF files.}
}
\value{
A list of paths to the extracted NetCDF files.
}
\description{
Unzip AGFD NetCDF files from ZIP
}
\keyword{internal}
