% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.abares-package.R
\name{.readabares_collaborators}
\alias{.readabares_collaborators}
\title{Read collaborators.txt}
\usage{
.readabares_collaborators()
}
\value{
a vector of GitHub usernames.
}
\description{
Reads a text file containing GitHub usernames of collaborators.
}
\author{
Adam H. Sparks and Maëlle Salmon
via (<https://github.com/EMODnet/emodnet.wfs/blob/69ca933e5a4154cb651b1d3158072f86d0a7ccb9/R/emodnet.wfs-package.R).
}
\keyword{internal}
