% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clum.R
\name{.set_clum_update_levels}
\alias{.set_clum_update_levels}
\title{Set CLUM scale update levels}
\usage{
.set_clum_update_levels()
}
\description{
Creates data.tables containing the raster categories for the \acronym{CLUM}
scale update levels.
}
\keyword{internal}
