% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.set_max_tries}
\alias{.set_max_tries}
\title{Set the Max Tries for Download Attempts}
\usage{
.set_max_tries()
}
\value{
An \code{integer} value to be used as the maximum number of tries to
retry when a download fails.
}
\description{
Allows users to specify a value via \code{options()} that overrides the default
value of \code{3L}.
}
\examples{
options(read.abares_max_tries = 4L)
.set_max_tries()

}
\keyword{internal}
