% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.set_timeout}
\alias{.set_timeout}
\title{Set the Timeout for Download Connections}
\usage{
.set_timeout()
}
\value{
An \code{integer} value to be used as the connection timeout value.
}
\description{
Allows users to specify a value via \code{options()} that overrides the default
value of \code{2000L}.
}
\examples{
options(read.abares_timeout = 4000L)
.set_timeout()

}
\keyword{internal}
