% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.set_ua}
\alias{.set_ua}
\title{Set the User Agent String}
\usage{
.set_ua()
}
\value{
A \code{string} value to be used as the user-agent in the HTTP request.
}
\description{
Allows users to specify a user-agent string via \code{options()} that overrides
the default string.
}
\examples{
options(read.abares_ua = "read.abares testing v0.0.0.9000")
.set_ua()

}
\keyword{internal}
