% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.try_resume}
\alias{.try_resume}
\title{Try to resume a partial download using brio}
\usage{
.try_resume(req, dest, existing_size)
}
\arguments{
\item{req}{httr2_request object.}

\item{dest}{Destination file path.}

\item{existing_size}{Size of existing partial file.}
}
\value{
Logical indicating if resume was successful.
}
\description{
Try to resume a partial download using brio
}
\keyword{internal}
