% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aagis_regions.R
\name{get_aagis_regions}
\alias{get_aagis_regions}
\title{Get AAGIS Region Mapping Files}
\usage{
get_aagis_regions(cache = TRUE)
}
\arguments{
\item{cache}{\code{Boolean} Cache the \acronym{AAGIS} regions shape files after
download using \code{tools::R_user_dir()} to identify the proper directory for
storing user data in a cache for this package. Defaults to \code{TRUE}, caching
the files locally as a native \R object. If \code{FALSE}, this function uses
\code{tempdir()} and the files are deleted upon closing of the \R session.}
}
\value{
An \CRANpkg{sf} object of the \acronym{AAGIS} regions
}
\description{
Get AAGIS Region Mapping Files
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aagis <- get_aagis_regions()

plot(aagis)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/research-topics/surveys/farm-definitions-methods#regions}
}
\seealso{
Other AGFD: 
\code{\link{get_agfd}()},
\code{\link{read_agfd_dt}()},
\code{\link{read_agfd_stars}()},
\code{\link{read_agfd_terra}()},
\code{\link{read_agfd_tidync}()}
}
\concept{AGFD}
