% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abares_trade.R
\name{get_abares_trade}
\alias{get_abares_trade}
\title{Get Data From the ABARES Trade Dashboard}
\usage{
get_abares_trade(cache = TRUE)
}
\arguments{
\item{cache}{\code{Boolean} Cache the \acronym{ABARES} trade data after download
using \code{tools::R_user_dir()} to identify the proper directory for storing
user data in a cache for this package. Defaults to \code{TRUE}, caching the files
locally as a native \R object. If \code{FALSE}, this function uses \code{tempdir()}
and the files are deleted upon closing of the \R session.}
}
\value{
A \CRANpkg{data.table} object of the \acronym{ABARES} trade data
}
\description{
Fetches and imports  \acronym{ABARES} trade data. Columns are renamed for
consistency using lower case, snake_case format.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
trade <- get_abares_trade()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0}
}
\seealso{
Other Trade: 
\code{\link{get_abares_trade_regions}()}
}
\concept{Trade}
