% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abares_trade_regions.R
\name{get_abares_trade_regions}
\alias{get_abares_trade_regions}
\title{Get ABARES Trade Data Regions From the ABARES Trade Dashboard}
\source{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/2}
}
\usage{
get_abares_trade_regions(cache = TRUE)
}
\arguments{
\item{cache}{\code{Boolean} Cache the \acronym{ABARES} trade regions data after
download using \code{tools::R_user_dir()} to identify the proper directory for
storing user data in a cache for this package. Defaults to \code{TRUE}, caching
the files locally as a native \R object. If \code{FALSE}, this function uses
\code{tempdir()} and the files are deleted upon closing of the \R session.}
}
\value{
A \CRANpkg{data.table} object of the \acronym{ABARES} trade data
regions.
}
\description{
Fetches and imports  \acronym{ABARES} trade regions data.
}
\note{
Columns are renamed for consistency with other \acronym{ABARES} products
serviced in this package using a snake_case format and ordered consistently.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
trade_regions <- get_abares_trade_regions()

trade_regions
\dontshow{\}) # examplesIf}
}
\seealso{
Other Trade: 
\code{\link{get_abares_trade}()}
}
\concept{Trade}
