% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agfd.R
\name{get_agdf}
\alias{get_agdf}
\title{Fetch Australian Gridded Farm Data}
\usage{
get_agdf(fixed = FALSE, cache = FALSE)
}
\arguments{
\item{fixed}{`Boolean` Download fixed historical climate prices?  Defaults
to `FALSE`, downloads the data from ABARES \sQuote{farmpredict} (Hughes
*et al.* 2022). Using `TRUE` will download simulations where global output
and input price indexes are fixed at values from the most recently completed
financial year.}

\item{cache}{`Boolean` Cache the Australian Gridded Farm Data files after
download using [tools::R_user_dir] to identify the proper directory for
storing user data in a cache for this package.}
}
\value{
A character `vector` of NetCDF files containing the Australian
 Gridded Farm Data with fullnames.
}
\description{
Fetch Australian Gridded Farm Data
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
agfd <- get_agdf()
\dontshow{\}) # examplesIf}
}
\references{
N. Hughes, W.Y. Soh, C. Boult, K. Lawson, Defining drought from
 the perspective of Australian farmers, Climate Risk Management, Volume 35,
 2022, 100420, ISSN 2212-0963, DOI
 [10.1016/j.crm.2022.100420](https://doi.org/10.1016/j.crm.2022.100420).
}
