% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_broadacre_crops_data.R
\name{get_broadacre_crops_data}
\alias{get_broadacre_crops_data}
\title{Get Broadacre Crops by Australia, State and Territory by Year}
\usage{
get_broadacre_crops_data(crops, year = "latest")
}
\arguments{
\item{crops}{A character vector providing the desired cropping data, one of:
\itemize{
\item winter,
\item summer or
\item sugarcane.
}}

\item{year}{A string value providing the year of interest to download.
Formatted as \code{"2022-23"} or \code{"2023-24"} or use \code{"latest"} for the most
recent release available. Defaults to \code{"latest"}.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} object of the requested data.
}
\description{
Technically these are from the Australian Bureau of Statistics (\acronym{ABS}
data, not \acronym{ABARES}, but the data is agricultural and so it's serviced
here.
}
\examples{
get_broadacre_crops_data()

}
