% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates_by_performance_category.R
\name{get_estimates_by_performance_category}
\alias{get_estimates_by_performance_category}
\alias{get_est_by_perf_cat}
\title{Get Estimates by Size From ABARES}
\source{
\url{https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download}
}
\usage{
get_estimates_by_performance_category()

get_est_by_perf_cat()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object
}
\description{
Get Estimates by Size From ABARES
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

 get_estimates_by_performance_category()

 # or shorter
 get_est_by_perf_cat()
\dontshow{\}) # examplesIf}
}
\seealso{
Other Estimates: 
\code{\link{get_estimates_by_size}()},
\code{\link{get_historical_national_estimates}()},
\code{\link{get_historical_regional_estimates}()},
\code{\link{get_historical_state_estimates}()}
}
\concept{Estimates}
