% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates_by_size.R
\name{get_estimates_by_size}
\alias{get_estimates_by_size}
\alias{get_est_by_size}
\title{Get Estimates by Size From ABARES}
\usage{
get_estimates_by_size()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object
}
\description{
Get Estimates by Size From ABARES
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_est_by_size()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_est_by_size()
\dontshow{\}) # examplesIf}
}
