% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_national_estimate.R
\name{get_historical_national_est}
\alias{get_historical_national_est}
\alias{get_hist_nat_est}
\title{Get Historical National Estimates from ABARES}
\usage{
get_historical_national_est()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object
}
\description{
Get Historical National Estimates from ABARES
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_hist_nat_est()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 get_hist_nat_est()
\dontshow{\}) # examplesIf}
}
