% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_national_estimates.R
\name{get_historical_national_estimates}
\alias{get_historical_national_estimates}
\alias{get_hist_nat_est}
\title{Get Historical National Estimates from ABARES}
\source{
\url{https://www.agriculture.gov.au/sites/default/files/documents/fdp-national-historical.csv}
}
\usage{
get_historical_national_estimates()

get_hist_nat_est()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object with the \code{Variable} field as the
\code{key}.
}
\description{
Get Historical National Estimates from ABARES
}
\note{
Columns are renamed for consistency with other \acronym{ABARES} products
serviced in this package using a snake_case format and ordered consistently.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

 get_historical_national_estimates()

 # or shorter
 get_hist_nat_est()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download}
}
\seealso{
Other Estimates: 
\code{\link{get_estimates_by_performance_category}()},
\code{\link{get_estimates_by_size}()},
\code{\link{get_historical_regional_estimates}()},
\code{\link{get_historical_state_estimates}()}
}
\concept{Estimates}
