% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_state_estimates.R
\name{get_historical_state_estimates}
\alias{get_historical_state_estimates}
\alias{get_hist_sta_est}
\title{Get Historical State Estimates from ABARES}
\usage{
get_historical_state_estimates()

get_hist_sta_est()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object
}
\description{
Get Historical State Estimates from ABARES
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 get_historical_state_estimates()

 # or shorter
 get_hist_sta_est()
\dontshow{\}) # examplesIf}
}
