% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topsoil_thickness.R
\name{get_topsoil_thickness}
\alias{get_topsoil_thickness}
\title{Get topsoil thickness for 'Australian Areas of Intensive Agriculture of Layer 1' for local use}
\source{
\url{https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip}
}
\usage{
get_topsoil_thickness(cache = FALSE)
}
\arguments{
\item{cache}{Caches the topsoil thickness data files after download
using \code{tools::R_user_dir()} to identify the proper directory for storing
user data in a cache for this package. Defaults to \code{FALSE}.
locally. If \code{FALSE}, this function uses \code{tempdir()} and the files are deleted
upon closing of the active \R session.}
}
\value{
A \code{read.abares.topsoil.thickness} object, which is a named \code{list()}
with the \code{\link[fs:path_file]{fs::path_file()}} of the data file and text file of metadata.
}
\description{
Get topsoil thickness for 'Australian Areas of Intensive Agriculture of Layer 1' for local use
}
\details{
The topsoil thickness data will be saved as a GeoTIFF file in the user's
cache if \code{cache = FALSE}.
}
\note{
A custom \code{print()} method is provided that will print the metadata associated
with these data. Examples are provided for interacting with the metadata
directly.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
x <- get_topsoil_thickness()

# View the metadata with pretty printing
x

# Extract the metadata as an object in your R session and use it with
# {pander}, useful for Markdown files

library(pander)
y <- x$metadata
pander(y)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
}
\seealso{
Other topsoil_thickness: 
\code{\link{print_topsoil_thickness_metadata}()},
\code{\link{read_topsoil_thickness_stars}()},
\code{\link{read_topsoil_thickness_terra}()}
}
\concept{topsoil_thickness}
