% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_winter_broadacre_crops.R
\name{get_winter_crops}
\alias{get_winter_crops}
\title{Get Winter Broadacre Crops by Australia, State and Territory by Year}
\usage{
get_winter_crops()
}
\value{
A [data.table::data.table() object of the requested data.
}
\description{
Get Winter Broadacre Crops by Australia, State and Territory by Year
}
\examples{
get_winter_crops()

}
