% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%||\%}
\alias{\%||\%}
\title{Utility operator for NULL coalescing}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value.}

\item{y}{Second value (used if x is \code{NULL}).}
}
\value{
\code{x} if not \code{NULL}, otherwise \code{y}.
}
\description{
Utility operator for NULL coalescing
}
\keyword{internal}
