% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_cache.R
\name{inspect_cache}
\alias{inspect_cache}
\title{List the File Path to Users' Cache Directory}
\usage{
inspect_cache(recursive = FALSE)
}
\arguments{
\item{recursive}{\code{Boolean} value indicating whether or not to list files in
subdirectories of the cache directory. Defaults to \code{FALSE} returning only
the top-level directories contained in the cache directory.}
}
\value{
A \code{list} of directories or files in the cache
}
\description{
Check what files exist in your \pkg{agfd} file cache.  This function will
always return full file names, \emph{i.e.}, the directory path is prepended.  See
the help file for \code{list.files()} for more on the \code{full.names} argument. If
you wish to strip the file path and only return the directory or filenames,
use \code{basename()}. See examples for more.
}
\examples{
# list directories in cache only
inspect_cache()

# list directory names, stripping the file path
basename(inspect_cache)

# list all files in subdirectories of the cache
inspect_cache(recursive = TRUE)

# list all files in subdirectories, stripping the file path
basename(inspect_cache(recursive_true))

}
\seealso{
Other cache: 
\code{\link{clear_cache}()}
}
\concept{cache}
