% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_abs_production_data.R
\name{parse_abs_production_data}
\alias{parse_abs_production_data}
\title{Create a Unified data.table Object of All Sheets in an Excel Workbook}
\usage{
parse_abs_production_data(filename)
}
\arguments{
\item{filename}{The name of the Excel workbook for import}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} object.
}
\description{
Create a Unified data.table Object of All Sheets in an Excel Workbook
}
\keyword{internal}
