% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_thickness.R
\name{print_soil_thickness_metadata}
\alias{print_soil_thickness_metadata}
\title{Display complete metadata associated with soil thickness data}
\usage{
print_soil_thickness_metadata(x)
}
\arguments{
\item{x}{A \code{read.abares.soil.thickness.files} object.}
}
\value{
Nothing, called for its side effects, it prints the complete
metadata file to the \R console.
}
\description{
Displays the complete set of metadata associated with the soil thickness
data in your \R console. For including the metadata in documents or other
methods outside of \R, see \code{\link[=get_soil_thickness]{get_soil_thickness()}} for an example using
\code{\link[pander:pander]{pander::pander()}} to print the metadata.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_soil_thickness(cache = TRUE) |>
  print_soil_thickness_metadata()
\dontshow{\}) # examplesIf}
}
\seealso{
Other soil_thickness: 
\code{\link{get_soil_thickness}()},
\code{\link{read_soil_thickness_stars}()},
\code{\link{read_soil_thickness_terra}()}
}
\concept{soil_thickness}
