% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.abares_options.R
\name{read.abares_options}
\alias{read.abares_options}
\title{Get or Set read.abares Options}
\usage{
read.abares_options(...)
}
\arguments{
\item{...}{Named options to set, or no arguments to retrieve current values.}
}
\value{
A list of current option values.
}
\description{
A convenience function to get or set options used by \pkg{read.abares}.
}
\examples{
# See currently set options for {read.abares}
read.abares_options()

# Set to "verbosity" to "quiet" suppress messages
read.abares_options(read.abares.verbosity = "quiet")
read.abares_options()

}
\seealso{
Other read.abares-options:
\code{\link{read.abares-options}}
}
\concept{read.abares-options}
