% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_broadacre_data.R
\name{read_abs_broadacre_data}
\alias{read_abs_broadacre_data}
\title{Get ABS' Broadacre Crops Production and Value by Australia, State and Territory by Year}
\usage{
read_abs_broadacre_data(crops = "winter", year = "latest", file = NULL)
}
\arguments{
\item{crops}{A character vector providing the desired cropping data, one of:
\itemize{
\item winter (default),
\item summer or
\item sugarcane.
}}

\item{year}{A string value providing the year of interest to download.
Formatted as \code{"2022-23"} or \code{"2023-24"} or use \code{"latest"} for the most
recent release available. Defaults to \code{"latest"}.}

\item{file}{A string value providing a file path to an \acronym{ABS}
Australian production data set that you have downloaded and saved locally.
No checks are performed, it will attempt to parse any file you pass to it
here. If this is used, other arguments will be ignored.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} object of the requested data.
}
\description{
Automates downloading and importing of \acronym{ABS} broadacre crop
production data. Please view the comments embedded in the spreadsheets
themselves (that really should be columns of comments on the data) for
important information. The \code{source} link provides a direct download for the
latest available spreadsheet to download and open in Excel.
}
\details{
Technically these data are from the Australian Bureau of Statistics
(\acronym{ABS}, not \acronym{ABARES}, but the data is agricultural and so
it's serviced in this package.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
read_abs_broadacre_data()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.abs.gov.au/statistics/industry/agriculture/australian-agriculture-broadacre-crops}.
}
\seealso{
Other ABS: 
\code{\link{read_abs_horticulture_data}()},
\code{\link{read_abs_livestock_data}()}
}
\concept{ABS}
