% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_livestock_data.R
\name{read_abs_livestock_data}
\alias{read_abs_livestock_data}
\title{Read Livestock Production and Value by Australia, State and Territory by Year}
\usage{
read_abs_livestock_data(year = "latest", file = NULL)
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} object of the requested data.
}
\description{
Automates downloading and importing of \acronym{ABS} livestock production
data. Please view the comments embedded in the spreadsheets themselves (that
really should be columns of comments on the data) for important information.
The \code{source} link provides a direct download for the spreadsheet to
download and open in Excel.
}
\details{
Technically these data are from the Australian Bureau of Statistics
(\acronym{ABS}, not \acronym{ABARES}, but the data is agricultural and so
it's serviced in this package.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
read_livestock_data()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.abs.gov.au/statistics/industry/agriculture/australian-agriculture-livestock}.
}
