% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_table.R, R/read_tibble.R
\name{read_agdf_dt}
\alias{read_agdf_dt}
\title{Read AGDF NCDF Files as a data.table}
\usage{
read_agdf_dt(files)

read_agdf_dt(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \code{list} of \CRANpkg{tidync} \link[tidync:tidync]{tidync::tidync} objects of the
Australian Gridded Farm Data

a \code{list} of \CRANpkg{tidync} \link[tidync:tidync]{tidync::tidync} objects of the
Australian Gridded Farm Data
}
\description{
Read AGDF NCDF Files as a data.table

Read AGDF NCDF Files as a tibble
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agdf_dt()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agdf_dt()
\dontshow{\}) # examplesIf}
}
\seealso{
Other read_agdf: 
\code{\link{read_agdf_stars}()},
\code{\link{read_agdf_terra}()},
\code{\link{read_agdf_tidync}()}

Other read_agdf: 
\code{\link{read_agdf_stars}()},
\code{\link{read_agdf_terra}()},
\code{\link{read_agdf_tidync}()}
}
\concept{read_agdf}
