% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agdf_stars.R
\name{read_agdf_stars}
\alias{read_agdf_stars}
\title{Read AGDF NCDF Files}
\usage{
read_agdf_stars(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}

\item{class}{Class of objec to return, `terra` or `stars`}
}
\description{
Read AGDF NCDF Files
}
