% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agdf_terra.R
\name{read_agdf_terra}
\alias{read_agdf_terra}
\title{Read AGDF NCDF Files}
\usage{
read_agdf_terra(class, files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \CRANpkg{terra} [terra::rast] object of the Australian Gridded Farm
 Data
}
\description{
Read AGDF NCDF Files
}
