% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agdf_tidync.R
\name{read_agdf_tidync}
\alias{read_agdf_tidync}
\title{Read AGDF NCDF Files with tidync}
\usage{
read_agdf_tidync(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a `list` of \CRANpkg{tidync} [tidync::tidync] objects of the
 Australian Gridded Farm Data
}
\description{
Read AGDF NCDF Files with tidync
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
terr_a <- get_agfd(cache = TRUE) |>
  read_agdf_terra()
\dontshow{\}) # examplesIf}
}
