% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agfd_dt.R
\name{read_agfd_dt}
\alias{read_agfd_dt}
\title{Read AGFD NCDF Files as a data.table}
\usage{
read_agfd_dt(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \link[data.table:data.table]{data.table::data.table} object of the Australian Gridded Farm Data
}
\description{
Read Australian Gridded Farm Data, (\acronym{AGFD}) as a \link{data.table} object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agfd_dt()
\dontshow{\}) # examplesIf}
}
\seealso{
Other AGFD: 
\code{\link{get_aagis_regions}()},
\code{\link{get_agfd}()},
\code{\link{read_agfd_stars}()},
\code{\link{read_agfd_terra}()},
\code{\link{read_agfd_tidync}()}
}
\concept{AGFD}
