% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_historical_forecast_database.R
\name{read_historical_forecast_database}
\alias{read_historical_forecast_database}
\alias{read_historical_forecast}
\title{Read ABARES' "Historical Forecast Database"}
\source{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1031941/0}
}
\usage{
read_historical_forecast_database(file = NULL)

read_historical_forecast(file = NULL)
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} object.
}
\description{
Fetches and imports \acronym{ABARES} "Historical Forecast Database"
performance data.
}
\note{
Columns are renamed for consistency with other \acronym{ABARES} products
serviced in this package using a snake_case format and ordered consistently.

The "Month_issued" column is converted from a character string to a numeric
value representing the month of year, \emph{e.g.}, "March" is converted to \code{3}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

read_historical_forecast_database()

# or shorter
read_historical_forecast()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/research-topics/agricultural-outlook/historical-forecasts#:~:text=About\%20the\%20historical\%20agricultural\%20forecast,relevant\%20to\%20Australian\%20agricultural\%20markets}
}
