% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abs_livestock_data.R
\name{read_livestock_data}
\alias{read_livestock_data}
\title{Read Livestock Production and Value by Australia, State and Territory by Year}
\usage{
read_livestock_data(year = "latest", file = NULL)
}
\arguments{
\item{year}{A string value providing the year of interest to download.
Formatted as \code{"2022-23"} or \code{"2023-24"} or use \code{"latest"} for the most
recent release available. Defaults to \code{"latest"}.}

\item{file}{A string value providing a file path to an \acronym{ABS}
Australian production data set that you have downloaded and saved locally.
No checks are performed, it will attempt to parse any file you pass to it
here. If this is used, the \code{crops} and \code{year} arguments will be ignored.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} object of the requested data.
}
\description{
Automates downloading and importing of \acronym{ABS} livestock production
data. Please view the comments embedded in the spreadsheets themselves (that
really should be columns of comments on the data) for important information.
The \code{source} link provides a direct download for the spreadsheet to
download and open in Excel.
}
\details{
Technically these data are from the Australian Bureau of Statistics
(\acronym{ABS}, not \acronym{ABARES}, but the data is agricultural and so
it's serviced in this package.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
read_livestock_data()
\dontshow{\}) # examplesIf}
}
