% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nlum_stars.R
\name{read_nlum_stars}
\alias{read_nlum_stars}
\title{Read national scale Land Use of Australia GeoTIFFs using stars}
\source{
\describe{
\item{Y201011}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_ALUMV8_2010_11_alb_package_20241128.zip}}
\item{Y201516}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_ALUMV8_2015_16_alb_package_20241128.zip}}
\item{Y202021}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_ALUMV8_2020_21_alb_package_20241128.zip}}
\item{C201021}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_CHANGE_SIMP_2011_to_2021_alb_package_20241128.zip}}
\item{T201011}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_INPUTS_2010_11_geo_package_20241128.zip}}
\item{T201516}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_INPUTS_2015_16_geo_package_20241128.zip}}
\item{T202021}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_INPUTS_2020_21_geo_package_20241128.zip}}
\item{P201011}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_AgProbabilitySurfaces_2010_11_geo_package_20241128.zip}}
\item{P201516}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_AgProbabilitySurfaces_2015_16_geo_package_20241128.zip}}
\item{P202021}{\url{https://www.agriculture.gov.au/sites/default/files/documents/NLUM_v7_250_AgProbabilitySurfaces_2020_21_geo_package_20241128.zip}}
}.
}
\usage{
read_nlum_stars(
  data_set = c("Y201011", "Y201516", "C201021", "T201011", "T201516", "T202021",
    "P201011", "P201516", "P202021"),
  cache = getOption("read.abares.cache"),
  user_agent = getOption("read.abares.user_agent"),
  max_tries = getOption("read.abares.max_tries"),
  timout = getOption("read.abares.max_tries"),
  ...
)
}
\arguments{
\item{data_set}{A string value indicating the GeoTIFF desired for download.
One of:
\describe{
\item{Y201011}{Land use of Australia 2010–11}
\item{Y201516}{Land use of Australia 2015–16}
\item{Y202021}{Land use of Australia 2020–21}
\item{C201021}{Land use of Australia change}
\item{T201011}{Land use of Australia 2010–11 thematic layers}
\item{T201516}{Land use of Australia 2015–16 thematic layers}
\item{T202021}{Land use of Australia 2020–21 thematic layers}
\item{P201011}{Land use of Australia 2010–11 agricultural commodities probability grids}
\item{P201516}{Land use of Australia 2015–16 agricultural commodities probability grids}
\item{P202021}{Land use of Australia 2020–21 agricultural commodities probability grids}
}.}

\item{cache}{Boolean cache the files after download? Defaults to \code{FALSE}
with files being downloaded to \code{tempdir()} being available throughout the
active \R session and cleaned up on exit.  If set to \code{TRUE}, files will be
cached locally for use between sessions.  See \dQuote{Caching} section for
more.}

\item{user_agent}{Character string providing a custom user agent to pass
along with the download request for the server logs.}

\item{max_tries}{Integer providing the number of times that a file download
should be retried upon failure.  Defaults to 3.}

\item{...}{Additional arguments passed to \link[stars:read_stars]{stars::read_stars}, for \emph{e.g.},
\code{RAT} if you wish to set the active category when loading any of the
available GeoTIFF files that are encoded with a raster attribute table.}

\item{active_cat}{A string value or integer indicating the active category to
be used for the raster. Note that this value is dependent upon the
\code{data_set} chosen and will always default to the first column of the
attribute table for the requested data.}
}
\value{
a \CRANpkg{stars} object that may be one or many layers depending
upon the requested data set.
}
\description{
Download and import national scale \dQuote{Land Use of Australia v7} GeoTIFFs
as a \CRANpkg{stars} object.  Data can be cached on request.
}
\details{
From the
\href{https://www.agriculture.gov.au/abares/aclump/land-use/land-use-of-australia-2010-11-to-2020-21}{ABARES website}:
\dQuote{The \emph{Land use of Australia 2010–11 to 2020–21} data package consists
of seamless continental rasters that present land use at national scale for
2010–11, 2015–16 and 2020–21 and the associated change between each target
period.  Non-agricultural land uses are mapped using 7 thematic layers,
derived from existing datasets provided by state and territory jurisdictions
and external agencies. These 7 layers are: protected areas, topographic
features, land tenure, forest type, catchment scale land use, urban
boundaries, and stock routes. The agricultural land uses are based on the
Australian Bureau of Statistics’ 2010–11, 2015–16 and 2020–21 agricultural
census data; with spatial distributions modelled using Terra Moderate
Resolution Imaging Spectroradiometer (\acronym{MODIS}) satellite imagery and
training data, assisted by spatial constraint layers for cultivation,
horticulture, and irrigation.
Land use is specified according to the Australian Land Use and Management
(\acronym{ALUM}) Classification version 8. The same method is applied to all
target periods using representative national datasets for each period, where
available. All rasters are in GeoTIFF format with geographic coordinates in
Geocentric Datum of Australian 1994 (GDA94) and a 0.002197 degree
(~250 metre) cell size.
The \emph{Land use of Australia 2010–11 to 2020–21} data package is a product
of the Australian Collaborative Land Use and Management Program. This data
package replaces the Land use of Australia 2010–11 to 2015–16 data package,
with updates to these time periods.}
-- \acronym{ABARES}, 2024-11-28
}
\note{
The raster will load with the default category for each data set, but you can
specify a different category to use by passing the \code{RAT} argument through
the \code{...}.  To see which categories are available, please refer
to the metadata for these data.  The \acronym{PDF} can be accessed in your
default \acronym{PDF} viewer by using \code{\link[=view_nlum_metadata_pdf]{view_nlum_metadata_pdf()}}.
}
\section{Caching}{


When caching is enabled, the directory defined by \code{\link[tools:userdir]{tools::R_user_dir()}} will
be used by default unless overidden using
\code{read.abares_options(cache_location = "your_desired_file_path")} to change
the file path used for caching. You may also set this globally to \code{TRUE} or
\code{FALSE} by using \code{read.abares_options(read.abares.cache = TRUE)} or
\code{read.abares_options("read.abares.cache" = TRUE)}, respectively. Using the
argument in this function will override any options that are set globally.
See \link{read.abares-options} for more.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

read_nlum_stars()

nlum_stars

plot(nlum_stars)
\dontshow{\}) # examplesIf}
}
\references{
ABARES 2024, Land use of Australia 2010–11 to 2020–21, Australian Bureau of
Agricultural and Resource Economics and Sciences, Canberra, November, CC BY
4.0. \doi{10.25814/w175-xh85}
}
\seealso{
Other nlum: 
\code{\link{read_nlum_terra}()},
\code{\link{view_clum_metadata_pdf}()},
\code{\link{view_nlum_metadata_pdf}()}
}
\concept{nlum}
