% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soil_thickness_stars.R
\name{read_soil_thickness_stars}
\alias{read_soil_thickness_stars}
\title{Read Soil Thickness File With stars}
\usage{
read_soil_thickness_stars(files)
}
\arguments{
\item{files}{An \pkg{abares} \code{abares.soil.thickness} object, a \code{list} that
contains the \acronym{ESRI} grid file to import}
}
\value{
a \link{stars} object of the Soil Thickness for Australian Areas of
Intensive Agriculture of Layer 1
}
\description{
Read Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
data as a \link{stars} object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_soil_thickness(cache = TRUE) |>
  read_soil_thickness_stars()
\dontshow{\}) # examplesIf}
}
\seealso{
Other soil_thickness: 
\code{\link{display_soil_thickness_metadata}()},
\code{\link{get_soil_thickness}()},
\code{\link{read_soil_thickness_terra}()}
}
\concept{soil_thickness}
