% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soil_thickness_terra.R
\name{read_soil_thickness_terra}
\alias{read_soil_thickness_terra}
\title{Read Soil Thickness File With terra}
\usage{
read_soil_thickness_terra(files)
}
\arguments{
\item{files}{An \pkg{read.abares} \code{read.abares.soil.thickness} object, a \code{list} that
contains the \acronym{ESRI} grid file to import}
}
\value{
a \link[terra:rast]{terra::rast} object of the Soil Thickness for Australian Areas of
Intensive Agriculture of Layer 1
}
\description{
Read Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
data as a \link[terra:rast]{terra::rast} object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- get_soil_thickness(cache = TRUE) |>
  read_soil_thickness_terra()

# terra::plot() is reexported for convience
plot(x)
\dontshow{\}) # examplesIf}
}
\seealso{
Other soil_thickness: 
\code{\link{display_soil_thickness_metadata}()},
\code{\link{get_soil_thickness}()},
\code{\link{read_soil_thickness_stars}()}
}
\concept{soil_thickness}
