% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.abares-package.R
\name{readabares_user_agent}
\alias{readabares_user_agent}
\title{Create a custom user-agent string}
\source{
\url{https://github.com/EMODnet/emodnet.wfs/blob/69ca933e5a4154cb651b1d3158072f86d0a7ccb9/R/emodnet.wfs-package.R}.
}
\usage{
readabares_user_agent()
}
\value{
A character string to be used by \CRANpkg{httr2} as a user-agent.
}
\description{
Creates the user-agent string for \pkg{read.abares} for default values based
on whether the use is CI/development or interactive use.
}
\author{
Adam H. Sparks and Maëlle Salmon
}
\keyword{internal}
