% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot}
\alias{activeCat}
\alias{levels}
\alias{categories}
\alias{cats}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{terra}{\code{\link[terra:activeCat]{activeCat()}}, \code{\link[terra:categories]{categories()}}, \code{\link[terra:cats]{cats()}}, \code{\link[terra:levels]{levels()}}, \code{\link[terra:plot]{plot()}}}
}}

