% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{set_curl_handle}
\alias{set_curl_handle}
\title{Create a curl Handle for read.abares to use}
\usage{
set_curl_handle()
}
\value{
A \link[curl:handle]{curl::handle} object with polite headers and options set.
}
\description{
Create a curl Handle for read.abares to use
}
\keyword{internal}
