% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_nlum_metadata_pdf.R
\name{view_nlum_metadata_pdf}
\alias{view_nlum_metadata_pdf}
\title{Displays PDF Metadata for ABARES' "National Land Use" (NLUM) Raster Files in a Native Viewer}
\usage{
view_nlum_metadata_pdf()
}
\value{
An invisible \code{NULL}. Called for its side-effects, opens the system's
native \acronym{PDF} viewer to display the requested metadata \acronym{PDF}
document.
}
\description{
Each National Land Use (\acronym{NLUM}) raster file comes with a
\acronym{PDF} of metadata. This function will open and display that file
using the native \acronym{PDF} viewer for any system with a graphical user
interface and \acronym{PDF} viewer configured.  If the file does not exist
locally, it will be fetched and displayed.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
view_nlum_metadata_pdf()
\dontshow{\}) # examplesIf}
}
\seealso{
Other nlum:
\code{\link[=read_nlum_stars]{read_nlum_stars()}},
\code{\link[=read_nlum_terra]{read_nlum_terra()}},
\code{\link[=view_clum_metadata_pdf]{view_clum_metadata_pdf()}}
}
\concept{nlum}
