% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nlum.R
\name{view_nlum_pdf}
\alias{view_nlum_pdf}
\title{Displays the PDF metadata for the National Land Use raster files in a native viewer}
\source{
\url{https://stackoverflow.com/a/65738472}
}
\usage{
view_nlum_pdf()
}
\value{
Called for its side-effects, opens the system's native PDF viewer to
display the requested metadata file.
}
\description{
Each National Land Use raster file comes with a PDF of the metadata. This
function will open and display that file using the native PDF viewer for any
system.
}
\examples{
view_nlum_pdf()
}
