% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Remove the user's cache directory and all cached files}
\usage{
clear_cache()
}
\value{
An invisible \code{NULL}, called for its side-effects, clearing the
cached files and deleting the cache directory.
}
\description{
Removes the cache directory and all files in the \pkg{read.abares} cache if
any exist locally.
}
\examples{
# not run because cached files shouldn't exist on CRAN or testing envs
\dontrun{
clear_cache()
}
}
\seealso{
Other cache: 
\code{\link{inspect_cache}()}
}
\concept{cache}
