% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry_download.R
\name{.apply_timeouts}
\alias{.apply_timeouts}
\title{Apply timeouts to an httr2 request (internal)}
\usage{
.apply_timeouts(req, t_o)
}
\arguments{
\item{req}{An \code{httr2_request} object.}

\item{t_o}{A list like that returned by \code{.dataset_timeouts()}.}
}
\value{
An \code{httr2_request} with timeouts applied.
}
\description{
Apply timeouts to an httr2 request (internal)
}
\keyword{internal}
