% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aagis_regions.R
\name{.download_aagis_shp}
\alias{.download_aagis_shp}
\title{Download the 'Australian Agricultural and Grazing Industries Survey' (AAGIS) regions shapefile}
\usage{
.download_aagis_shp(.cache, .talktalk)
}
\arguments{
\item{.cache}{Boolean, enable caching?}

\item{.talktalk}{Boolean, how verbose should the function be?}
}
\value{
An \CRANpkg{sf} object of AAGIS regions.
}
\description{
Handles downloading, caching (if requested) and importing of AAGIS regions
geospatial data.  The geometries are corrected for validity before returning
to the user.
}
\keyword{internal}
