% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.find_user_cache}
\alias{.find_user_cache}
\title{Find the file path to the users' cache directory}
\usage{
.find_user_cache()
}
\value{
A \code{character} string value of a file path indicating the proper
directory to use for cached files.
}
\description{
Find the file path to the users' cache directory
}
\keyword{internal}
