% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clum.R
\name{.get_clum}
\alias{.get_clum}
\title{Get catchment scale Land Use of Australia data for local use}
\source{
\url{https://10.25814/2w2p-ph98}.
}
\usage{
.get_clum(.cache)
}
\arguments{
\item{.data_set}{A string value indicating the data desired for download.
One of:
\describe{
\item{CLUM_50m_2023v2}{Catchment Scale Land Use of Australia – Update December 2023 version 2}
\item{date_CLUM2023}{Catchment Scale Land Use of Australia - Date and Scale of Mapping}
\item{CLUM_Commodities_2023}{Catchment Scale Land Use of Australia – Commodities – Update December 2023}
}}
}
\value{
A \code{read.abares.clum} object, a list of files containing a spatial
data file of or related to Australian catchment scale land use data.
}
\description{
An internal function used by \link{read_clum_terra} and \link{read_clum_stars} that
downloads national level land use data GeoTIFF file, unzips the download
file and deletes unnecessary files that are included in the download.  Data
are cached on request.
}
\details{
The \code{CLUM_50m_2023v2} and \code{date_CLUM2023} datasets are available as GeoTIFF
files. The \code{CLUM_Commodities_2023} dataset is available as a shapefile.
The GeoTIFF files are both saved in the same format. The
\code{CLUM_Commodities_2023} file is saved as a GeoPackage after correcting
invalid geometries.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
CLUM50m <- get_clum(data_set = "CLUM_50m_2023v2")

CLUM50m
\dontshow{\}) # examplesIf}
}
\references{
ABARES 2024, Catchment Scale Land Use of Australia – Update December 2023
version 2, Australian Bureau of Agricultural and Resource Economics and
Sciences, Canberra, June, CC BY 4.0, DOI: \doi{10.25814/2w2p-ph98}.
}
\seealso{
Other Land Use: 
\code{\link{.get_nlum}()}
}
\concept{Land Use}
\keyword{internal}
