% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nlum.R
\name{.get_nlum}
\alias{.get_nlum}
\title{Get ABARES' National Scale "Land Use of Australia" Data}
\usage{
.get_nlum(.data_set, .x)
}
\arguments{
\item{.data_set}{A string value indicating the GeoTIFF desired for download.
One of:
\describe{
\item{Y201011}{Land use of Australia 2010–11}
\item{Y201516}{Land use of Australia 2015–16}
\item{Y202021}{Land use of Australia 2020–21}
\item{C201121}{Land use of Australia change}
\item{T201011}{Land use of Australia 2010–11 thematic layers}
\item{T201516}{Land use of Australia 2015–16 thematic layers}
\item{T202021}{Land use of Australia 2020–21 thematic layers}
\item{P201011}{Land use of Australia 2010–11 agricultural commodities probability grids}
\item{P201516}{Land use of Australia 2015–16 agricultural commodities probability grids}
\item{P202021}{Land use of Australia 2020–21 agricultural commodities probability grids}
}.}
}
\value{
An invisible \code{NULL} called for its side effect of downloading the
desired file.
}
\description{
An internal function used by \code{\link[=read_nlum_terra]{read_nlum_terra()}} and \code{\link[=read_nlum_stars]{read_nlum_stars()}} that
downloads national level land use data GeoTIFF file.
}
\examples{
.get_nlum(.data_set = "Y202021")

}
\references{
ABARES 2024, Land use of Australia 2010–11 to 2020–21, Australian Bureau of
Agricultural and Resource Economics and Sciences, Canberra, November, CC BY
4.0. \doi{10.25814/w175-xh85}.
}
\keyword{internal}
