% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nlum.R
\name{.get_nlum}
\alias{.get_nlum}
\title{Get national scale Land Use of Australia data for local use}
\source{
\url{https://doi.org/10.25814/w175-xh85}
}
\usage{
.get_nlum(.data_set, .cache, .user_agent, .max_tries, .timeout)
}
\arguments{
\item{.data_set}{A string value indicating the GeoTIFF desired for download.
One of:
\describe{
\item{Y201011}{Land use of Australia 2010–11}
\item{Y201516}{Land use of Australia 2015–16}
\item{Y202021}{Land use of Australia 2020–21}
\item{C201021}{Land use of Australia change}
\item{T201011}{Land use of Australia 2010–11 thematic layers}
\item{T201516}{Land use of Australia 2015–16 thematic layers}
\item{T202021}{Land use of Australia 2020–21 thematic layers}
\item{P201011}{Land use of Australia 2010–11 agricultural commodities probability grids}
\item{P201516}{Land use of Australia 2015–16 agricultural commodities probability grids}
\item{P202021}{Land use of Australia 2020–21 agricultural commodities probability grids}
}.}

\item{.cache}{Boolean cache the files after download? Defaults to \code{FALSE}
with files being downloaded to \code{tempdir()} being available throughout the
active \R session and cleaned up on exit.  If set to \code{TRUE}, files will be
cached locally for use between sessions.  See \dQuote{Caching} section for
more.}

\item{.user_agent}{Character string providing a custom user agent to pass
along with the download request for the server logs.}

\item{.max_tries}{Integer providing the number of times that a file download
should be retried upon failure.  Defaults to 3.}

\item{.timeout}{Integer value providing the number of seconds to wait before
timing out a download request.  Defaults to 2000.}
}
\value{
A \code{read.abares.nlum} object, a list of GeoTIFF files containing
national scale land use data and a PDF file of metadata.
}
\description{
An internal function used by \link{read_nlum_terra} and \link{read_nlum_stars} that
downloads national level land use data GeoTIFF file, unzips the download
file and deletes unnecessary files that are included in the download.  Data
are cached on request.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
Y202021 <- get_nlum(data_set = "Y202021")

Y202021
\dontshow{\}) # examplesIf}
}
\references{
ABARES 2024, Land use of Australia 2010–11 to 2020–21, Australian Bureau of
Agricultural and Resource Economics and Sciences, Canberra, November, CC BY
4.0. \doi{10.25814/w175-xh85}.
}
\keyword{internal}
