% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nlum.R
\name{.get_nlum}
\alias{.get_nlum}
\title{Get national scale Land Use of Australia data for local use}
\source{
\url{https://doi.org/10.25814/w175-xh85}
}
\usage{
.get_nlum(.data_set, .cache)
}
\arguments{
\item{.data_set}{A string value indicating the GeoTIFF desired for download.
One of:
\describe{
\item{Y201011}{Land use of Australia 2010–11}
\item{Y201516}{Land use of Australia 2015–16}
\item{Y202021}{Land use of Australia 2020–21}
\item{C201021}{Land use of Australia change}
\item{T201011}{Land use of Australia 2010–11 thematic layers}
\item{T201516}{Land use of Australia 2015–16 thematic layers}
\item{T202021}{Land use of Australia 2020–21 thematic layers}
\item{P201011}{Land use of Australia 2010–11 agricultural commodities probability grids}
\item{P201516}{Land use of Australia 2015–16 agricultural commodities probability grids}
\item{P202021}{Land use of Australia 2020–21 agricultural commodities probability grids}
}}

\item{.cache}{Cache the Australian Gridded Farm Data files after download
using \link[tools:userdir]{tools::R_user_dir} to identify the proper directory for storing user
data in a cache for this package. Defaults to \code{TRUE}, caching the files
locally. If \code{FALSE}, this function uses \code{tempdir()} and the files are
deleted upon closing of the active \R session.}
}
\value{
A \code{read.abares.nlum} object, a list of files containing a GeoTIFF of
national scale land use data and a PDF file of metadata.
}
\description{
An internal function used by \link{read_nlum_terra} and \link{read_nlum_stars} that
downloads national level land use data GeoTIFF file, unzips the download
file and deletes unnecessary files that are included in the download.  Data
are cached on request.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
Y202021 <- get_nlum(data_set = "Y202021")

Y202021
\dontshow{\}) # examplesIf}
}
\references{
ABARES 2024, Land use of Australia 2010–11 to 2020–21, Australian Bureau of
Agricultural and Resource Economics and Sciences, Canberra, November, CC BY
4.0. \doi{10.25814/w175-xh85}
}
\concept{Land Use}
\keyword{internal}
