% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topsoil_thickness.R
\name{.get_topsoil_thickness}
\alias{.get_topsoil_thickness}
\title{Get topsoil thickness for \dQuote{Australian Areas of Intensive Agriculture of Layer 1} for local use}
\source{
\url{https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip}
}
\usage{
.get_topsoil_thickness(.files = NULL)
}
\arguments{
\item{.files}{A character string passed that provides a file path to the
local directory holding the unzipped files for topsoil thickness.}
}
\value{
A \code{read.abares.topsoil.thickness} object, which is a named \code{list()}
with the \code{\link[terra:rast]{terra::rast()}} object of the data and text file of metadata.
}
\description{
Fetches topsoil thickness data and associated metadata from \acronym{ABARES}.
}
\note{
A custom \code{print()} method is provided that will print the metadata associated
with these data. Examples are provided for interacting with the metadata
directly.
}
\examples{
x <- .read_topsoil_thickness()

# View the metadata with pretty printing
x

# Extract the metadata as an object in your R session and use it with
# {pander}, useful for Markdown files

library(pander)
y <- x$metadata
pander(y)

}
\references{
\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
}
\seealso{
Other topsoil thickness: 
\code{\link{print_topsoil_thickness_metadata}()},
\code{\link{read_topsoil_thickness_stars}()},
\code{\link{read_topsoil_thickness_terra}()}
}
\concept{topsoil thickness}
\keyword{internal}
