% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.retry_download}
\alias{.retry_download}
\title{Use httr2 to fetch a file with retries}
\usage{
.retry_download(url, .f)
}
\arguments{
\item{url}{Character the URL being requested.}

\item{.f}{Character` a filepath to be written to local storage.}

\item{.max_tries}{Integer the number of times to retry downloading the file
before failing.}

\item{.user_agent}{Character a string value with a custom user-defined
user-agent.}

\item{.timout}{Integer maximum number of seconds to wait before timing out
the request.}
}
\value{
An invisible \code{NULL}, called for its side-effects, writes an object
to the \code{tempdir()} for reading into the active \R session later.
}
\description{
Retries to download the requested resource before stopping. Uses
\CRANpkg{httr2} to cache in-session results in the \code{tempdir()}.
}
\examples{

f <- fs::path(tempdir(), "fdp-beta-national-historical.csv")
.retry_download(
  url = "https://www.agriculture.gov.au/sites/default/files/documents/fdp-beta-national-historical.csv",
  .f = f
)

}
\keyword{internal}
