% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clum_terra.R
\name{.set_clum_update_levels}
\alias{.set_clum_update_levels}
\title{Set CLUM Scale Update Levels}
\usage{
.set_clum_update_levels()
}
\description{
Creates data.tables containing the raster categories for the \acronym{CLUM}
scale update levels.
}
\keyword{internal}
