% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.unzip_file}
\alias{.unzip_file}
\title{Unzip a zip file}
\usage{
.unzip_file(.x)
}
\arguments{
\item{.x}{A zip file for unzipping.}
}
\value{
Invisible directory path, called for side effects.
}
\description{
Unzips the provided zip file into a folder named after the zip (without
.zip).
If unzipping fails, it will return an error and delete any partially-created
extract directory.
}
\keyword{internal}
