% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_file.R
\name{.unzip_file}
\alias{.unzip_file}
\title{Unzip a Zip File}
\usage{
.unzip_file(.file)
}
\arguments{
\item{.file}{A zip file for unzipping}
}
\value{
Called for its side-effects of unzipping a file, returns an
invisible \code{NULL}.
}
\description{
Unzips the provided zip file. If unzipping fails, it will return an error and
delete the corrupted zip file.
}
\keyword{internal}
