% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agfd.R
\name{get_agfd}
\alias{get_agfd}
\title{Fetch Australian Gridded Farm Data}
\usage{
get_agfd(fixed = FALSE, cache = TRUE)
}
\arguments{
\item{fixed}{\code{Boolean} Download fixed historical climate prices?  Defaults
to \code{FALSE}, downloads the data from ABARES \sQuote{farmpredict} (Hughes
\emph{et al.} 2022). Using \code{TRUE} will download simulations where global output
and input price indexes are fixed at values from the most recently completed
financial year.}

\item{cache}{\code{Boolean} Cache the Australian Gridded Farm Data files after
download using \link[tools:userdir]{tools::R_user_dir} to identify the proper directory for
storing user data in a cache for this package. Defaults to \code{TRUE}, caching
the files locally. If \code{FALSE}, this function uses \code{tempdir()} and the files
are deleted upon closing of the \R session.}
}
\value{
A character \code{vector} of NetCDF files containing the Australian
Gridded Farm Data with fullnames.
}
\description{
Fetch Australian Gridded Farm Data
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd()
\dontshow{\}) # examplesIf}
}
\references{
N. Hughes, W.Y. Soh, C. Boult, K. Lawson, Defining drought from
the perspective of Australian farmers, Climate Risk Management, Volume 35,
2022, 100420, ISSN 2212-0963, DOI
\href{https://doi.org/10.1016/j.crm.2022.100420}{10.1016/j.crm.2022.100420}.
}
