% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agfd.R
\name{get_agfd}
\alias{get_agfd}
\title{Fetch Australian Gridded Farm Data}
\usage{
get_agfd(fixed_prices = TRUE, cache = TRUE)
}
\arguments{
\item{fixed_prices}{\code{Boolean} Download historical climate and prices or
historical climate and fixed prices as described in  (Hughes \emph{et al.} 2022).
Defaults to \code{TRUE} and downloads the data with historical climate and fixed
prices \dQuote{to isolate the effects of climate variability on financial
incomes for broadacre farm businesses} (ABARES 2024). Using \code{TRUE} will
download simulations where global output and input price indexes are fixed
at values from the most recently completed financial year.}

\item{cache}{\code{Boolean} Cache the Australian Gridded Farm Data files after
download using \link[tools:userdir]{tools::R_user_dir} to identify the proper directory for
storing user data in a cache for this package. Defaults to \code{TRUE}, caching
the files locally. If \code{FALSE}, this function uses \code{tempdir()} and the files
are deleted upon closing of the \R session.}
}
\value{
A \code{read.abares.agfd.nc.files} object, a \code{list} of NetCDF files
containing the Australian Gridded Farm Data
}
\description{
Downloads The Australian Gridded Farm Data (\acronym{AGFD}) data and unzips
the compressed files to NetCDF for importing.
}
\details{
From the \href{https://www.agriculture.gov.au/abares/research-topics/surveys/farm-survey-data/australian-gridded-farm-data}{ABARES website}:
\dQuote{The Australian Gridded Farm Data (\acronym{AGFD}) are a set of national
scale maps containing simulated data on historical broadacre farm business
outcomes including farm profitability on an 0.05-degree (approximately 5 km)
grid.\cr
These data have been produced by \acronym{ABARES} as part of the ongoing
Australian Agricultural Drought Indicator (\acronym{AADI}) project
(previously known as the Drought Early Warning System Project) and were
derived using \acronym{ABARES}
\href{https://www.agriculture.gov.au/abares/research-topics/climate/drought/farmpredict}{\emph{farmpredict}}
model, which in turn is based on ABARES Agricultural and Grazing Industries
Survey (\acronym{AAGIS}) data.\cr
\href{https://www.agriculture.gov.au/abares/research-topics/climate/australian-agriculture-drought-indicators-project}{Australian Agricultural Drought Indicator}
(\acronym{AADI}) project (previously known as the Drought Early Warning
System Project) and were derived using \acronym{ABARES}
\href{https://www.agriculture.gov.au/abares/research-topics/climate/drought/farmpredict}{\emph{farmpredict}}
model, which in turn is based on ABARES \href{https://www.agriculture.gov.au/abares/research-topics/surveys/farm-definitions-methods}{Agricultural and Grazing Industries Survey}
(\acronym{AAGIS}) data.\cr
These maps provide estimates of farm business profit, revenue, costs and
production by location (grid cell) and year for the period 1990-91 to
2022-23. The data do not include actual observed outcomes but rather model
predicted outcomes for representative or \sQuote{typical} broadacre farm
businesses at each location considering likely farm characteristics and
prevailing weather conditions and commodity prices.}\cr
-- \acronym{ABARES}, 2024-11-25

Both sets of data are large in file size, \emph{i.e.}, >1GB, and will require time
to download.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd()
\dontshow{\}) # examplesIf}
}
\references{
\emph{Australian gridded farm data}, Australian Bureau of Agricultural and
Resource Economics and Sciences, Canberra, July 2024, DOI:
\href{https://doi.org/10.25814/7n6z-ev41}{10.25814/7n6z-ev41}.
\href{https://creativecommons.org/licenses/by/4.0/legalcode}{CC BY 4.0}.

N. Hughes, W.Y. Soh, C. Boult, K. Lawson, \emph{Defining drought from
the perspective of Australian farmers}, Climate Risk Management, Volume 35,
2022, 100420, ISSN 2212-0963, DOI:
\href{https://doi.org/10.1016/j.crm.2022.100420}{10.1016/j.crm.2022.100420}.

@source
\itemize{
\item Historical climate prices fixed -- \url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1036161/3},
\item Historical climate and prices -- \url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1036161/2}
}
}
\seealso{
Other AGFD: 
\code{\link{get_aagis_regions}()},
\code{\link{read_agfd_dt}()},
\code{\link{read_agfd_stars}()},
\code{\link{read_agfd_terra}()},
\code{\link{read_agfd_tidync}()}
}
\concept{AGFD}
