% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agfd.R
\name{get_agfd}
\alias{get_agfd}
\title{Fetch Australian Gridded Farm Data}
\usage{
get_agfd(fixed_prices = TRUE, cache = TRUE)
}
\arguments{
\item{fixed_prices}{\code{Boolean} Download historical climate and prices or
historical climate and fixed prices as described in  (Hughes \emph{et al.} 2022).
Defaults to \code{TRUE} and downloads the data with historical climate and fixed
prices \dQuote{to isolate the effects of climate variability on financial
incomes for broadacre farm businesses} (ABARES 2024). Using \code{TRUE} will
download simulations where global output and input price indexes are fixed
at values from the most recently completed financial year.}

\item{cache}{\code{Boolean} Cache the Australian Gridded Farm Data files after
download using \link[tools:userdir]{tools::R_user_dir} to identify the proper directory for
storing user data in a cache for this package. Defaults to \code{TRUE}, caching
the files locally. If \code{FALSE}, this function uses \code{tempdir()} and the files
are deleted upon closing of the \R session.}
}
\value{
A \code{list} of NetCDF files containing the Australian Gridded Farm Data
with fullnames, \emph{i.e.}, with the file path prepended to the file name
}
\description{
Fetch Australian Gridded Farm Data
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd()
\dontshow{\}) # examplesIf}
}
\references{
\emph{Australian gridded farm data}, Australian Bureau of Agricultural and
Resource Economics and Sciences, Canberra, July, DOI:
\href{https://doi.org/10.25814/7n6z-ev41}{10.25814/7n6z-ev41}. CC BY 4.0.

N. Hughes, W.Y. Soh, C. Boult, K. Lawson, \emph{Defining drought from
the perspective of Australian farmers}, Climate Risk Management, Volume 35,
2022, 100420, ISSN 2212-0963, DOI:
\href{https://doi.org/10.1016/j.crm.2022.100420}{10.1016/j.crm.2022.100420}.
}
