% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates_by_size.R
\name{get_estimates_by_size}
\alias{get_estimates_by_size}
\alias{get_est_by_size}
\title{Get Estimates by Size From ABARES}
\source{
\url{https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download}
}
\usage{
get_estimates_by_size()

get_est_by_size()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object with the \code{Variable} field as the
\code{key}.
}
\description{
Get Estimates by Size From ABARES
}
\note{
Columns are renamed and reordered for consistency.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

get_estimates_by_size()

# or shorter
get_est_by_size()
\dontshow{\}) # examplesIf}
}
\seealso{
Other Estimates: 
\code{\link{get_estimates_by_performance_category}()},
\code{\link{get_historical_national_estimates}()},
\code{\link{get_historical_state_estimates}()}
}
\concept{Estimates}
