% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_regional_estimates.R
\name{get_historical_regional_estimates}
\alias{get_historical_regional_estimates}
\alias{get_hist_reg_est}
\title{Get Historical Regional Estimates from ABARES}
\usage{
get_historical_regional_estimates()

get_hist_reg_est()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object with the \code{variable} field as the
\code{key}.
}
\description{
Get Historical Regional Estimates from ABARES
}
\note{
Columns are renamed and reordered for consistency.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 get_historical_regional_estimates()

 # or shorter
 get_hist_reg_est()
\dontshow{\}) # examplesIf}
}
