% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_thickness.R
\name{get_soil_thickness}
\alias{get_soil_thickness}
\title{Download Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1}
\usage{
get_soil_thickness(cache = TRUE)
}
\arguments{
\item{cache}{\code{Boolean} Cache the soil thickness data files after download
using \link[tools:userdir]{tools::R_user_dir} to identify the proper directory for storing user
data in a cache for this package. Defaults to \code{TRUE}, caching the files
locally. If \code{FALSE}, this function uses \code{tempdir()} and the files are deleted
upon closing of the \R session.}
}
\value{
A named \code{list} object with the file path of the resulting
\acronym{ESRI} Grid file and text file of metadata
}
\description{
Download Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_soil_thickness()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
}
