% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_thickness.R
\name{get_soil_thickness}
\alias{get_soil_thickness}
\title{Get soil thickness for 'Australian Areas of Intensive Agriculture of Layer 1' for local use}
\source{
\url{https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip}
}
\usage{
get_soil_thickness(cache = TRUE)
}
\arguments{
\item{cache}{Caches the soil thickness data files after download
using \code{tools::R_user_dir()} to identify the proper directory for storing
user data in a cache for this package. Defaults to \code{TRUE}, caching the files
locally. If \code{FALSE}, this function uses \code{tempdir()} and the files are deleted
upon closing of the active \R session.

A custom \code{print} method is provided that will print the metadata associated
with these data. Examples are provided for interacting with the metadata
directly.}
}
\value{
A \code{read.abares.soil.thickness} object, which is a named \code{list} with
the file path of the resulting Esri Grid file and text file of metadata.
}
\description{
Get soil thickness for 'Australian Areas of Intensive Agriculture of Layer 1' for local use
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- get_soil_thickness()

# View the metadata with pretty printing
x

# Extract the metadata as an object in your R session and use it with
# {pander}, useful for Markdown files

library(pander)
y <- x$metadata
pander(y)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
}
\seealso{
Other soil_thickness: 
\code{\link{print_soil_thickness_metadata}()},
\code{\link{read_soil_thickness_stars}()},
\code{\link{read_soil_thickness_terra}()}
}
\concept{soil_thickness}
