% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topsoil_thickness.R
\name{get_topsoil_thickness}
\alias{get_topsoil_thickness}
\title{Get topsoil thickness for 'Australian Areas of Intensive Agriculture of Layer 1' for local use}
\source{
\url{https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip}
}
\usage{
get_topsoil_thickness(cache = getOption("read.abares.cache"))
}
\arguments{
\item{cache}{Boolean cache the files after download? Defaults to \code{FALSE}
with files being downloaded to \code{tempdir()} being available throughout the
active \R session and cleaned up on exit.  If set to \code{TRUE}, files will be
cached locally for use between sessions.  See \dQuote{Caching} section for
more.}
}
\value{
A \code{read.abares.topsoil.thickness} object, which is a named \code{list()}
with the \code{\link[fs:path_file]{fs::path_file()}} of the data file and text file of metadata.
}
\description{
Fetches topsoil thickness data and associated metadata from \acronym{ABARES}.
}
\note{
A custom \code{print()} method is provided that will print the metadata associated
with these data. Examples are provided for interacting with the metadata
directly.
}
\section{Caching}{


When caching is enabled, the directory defined by \code{\link[tools:userdir]{tools::R_user_dir()}} will
be used by default unless overidden using
\code{read.abares_options(cache_location = "your_desired_file_path")} to change
the file path used for caching. You may also set this globally to \code{TRUE} or
\code{FALSE} by using \code{read.abares_options(read.abares.cache = TRUE)} or
\code{read.abares_options("read.abares.cache" = TRUE)}, respectively. Using the
argument in this function will override any options that are set globally.
See \code{\link[=read.abares-options]{read.abares-options()}} for more.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
x <- get_topsoil_thickness()

# View the metadata with pretty printing
x

# Extract the metadata as an object in your R session and use it with
# {pander}, useful for Markdown files

library(pander)
y <- x$metadata
pander(y)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
}
\seealso{
Other topsoil thickness: 
\code{\link{print_topsoil_thickness_metadata}()},
\code{\link{read_topsoil_thickness_stars}()},
\code{\link{read_topsoil_thickness_terra}()}
}
\concept{topsoil thickness}
