% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%||\%}
\alias{\%||\%}
\title{Return the left-hand side if it's not NULL, otherwise it return the right-hand side}
\usage{
x \%||\% y
}
\description{
Return the left-hand side if it's not NULL, otherwise it return the right-hand side
}
\keyword{internal}
