% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_cache.R
\name{inspect_cache}
\alias{inspect_cache}
\title{List the directories and files in your cache directory}
\usage{
inspect_cache(recurse = FALSE)
}
\arguments{
\item{recurse}{Boolean value indicating whether or not to \code{\link[fs:dir_ls]{fs::dir_ls()}} in
subdirectories of the cache directory. Defaults to \code{FALSE} returning only
the top-level directories and files contained in the cache directory.}
}
\value{
An \code{\link[fs:path]{fs::path()}} object containing directories or files in the cache.
}
\description{
Check what files exist in your \pkg{read.abares} file cache.  This function
will always return full file names, \emph{i.e.}, the directory path is prepended
by default.  See the help file for \code{\link[fs:dir_ls]{fs::dir_ls()}} for more on returning full
paths.  If you wish to strip the full path and only return the directory or
file names, see \code{\link[fs:path_file]{fs::path_file()}} or \code{\link[fs:path_file]{fs::path_dir()}}.
}
\examples{
# not run because cached files shouldn't exist on CRAN or testing envs
\dontrun{
library(fs)
# list directories in cache only
inspect_cache()

# list all files in subdirectories of the cache
inspect_cache(recurse = TRUE)

# list directory names, stripping the fs::path_file
path_file(inspect_cache)

# list all files in subdirectories, stripping the fs::path_file
path_file(inspect_cache(recurse = TRUE))
}
}
\seealso{
Other cache: 
\code{\link{clear_cache}()}
}
\concept{cache}
