% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topsoil_thickness.R
\name{print_topsoil_thickness_metadata}
\alias{print_topsoil_thickness_metadata}
\title{Display complete metadata associated with soil thickness data}
\usage{
print_topsoil_thickness_metadata(x)
}
\arguments{
\item{x}{A \code{read.abares.topsoil.thickness.files} object.}
}
\value{
Nothing, called for its side effects, it prints the complete
metadata file to the \R console.
}
\description{
Displays the complete set of metadata associated with the soil thickness
data in your \R console. For including the metadata in documents or other
methods outside of \R, see \code{get_topsoil_thickness()} for an example using
\code{\link[pander:pander]{pander::pander()}} to print the metadata.
}
\note{
The original metadata use a title of "Soil Thickness", in the context of this
package, we refer to it as "Topsoil Thickness" to be consistent with the
actual values in the data.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
get_topsoil_thickness() |>
  print_topsoil thickness_metadata()
\dontshow{\}) # examplesIf}
}
\seealso{
Other topsoil thickness: 
\code{\link{.get_topsoil_thickness}()},
\code{\link{read_topsoil_thickness_stars}()},
\code{\link{read_topsoil_thickness_terra}()}
}
\concept{topsoil thickness}
