% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.abares-options.R
\name{read.abares-options}
\alias{read.abares-options}
\title{Package Options for read.abares}
\description{
This page documents the global options used by the \pkg{read.abares} package.
}
\details{
The following options can be set via \code{\link[=options]{options()}} to control package behavior:

\describe{
   \item{\code{read.abares.ua}}{String. Set a custom useragent for web requests. Default is \code{"read.abares"}.}
   \item{\code{myPackage.timeout}}{Numeric. Timeout in seconds for operations. Default is \code{30}.}
}

These options can be set globally using:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(myPackage.ua = "myCustomUserAgent")
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=options]{options()}}, \code{\link[=getOption]{getOption()}}
}
\keyword{internal}
