% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.abares-package.R
\docType{package}
\name{read.abares-package}
\alias{read.abares}
\alias{read.abares-package}
\title{read.abares: Provides simple downloading, parsing and importing of Australian Bureau of Agricultural and Resource Economics and Sciences (ABARES) data sources}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Download and import data from the Australian Bureau of Agricultural and Resource Economics and Sciences (ABARES) \url{https://www.agriculture.gov.au/abares}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adamhsparks/read.abares}
  \item \url{https://adamhsparks.github.io/read.abares/}
  \item Report bugs at \url{https://github.com/adamhsparks/read.abares/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Other contributors:
\itemize{
  \item Jacob \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID}) (Assisted with troubleshooting formatting in documentation to display '<' and '>' properly) [contributor]
  \item Curtin University of Technology (Provided support through Adam Sparks's time.) [copyright holder]
  \item Grains Research and Development Corporation (GRDC Project CUR2210-005OPX (AAGI-CU)) [funder, copyright holder]
}

}
\keyword{internal}
