% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.abares-package.R
\docType{package}
\name{read.abares-package}
\alias{read.abares}
\alias{read.abares-package}
\title{read.abares: Read Australian Agricultural Data from Government Agencies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Downloads and imports agricultural data from the Australian Bureau of Agricultural and Resource Economics and Sciences (ABARES) \url{https://www.agriculture.gov.au/abares} and the Australian Bureau of Statistics (ABS) \url{https://www.abs.gov.au}. Supports multiple data formats including spreadsheets, comma‑separated value (CSV) files, and geospatial data such as shapefiles and GeoTIFFs. Covers topics such as broadacre crops, livestock, soils, commodities and related agricultural information. The package standardises field names and data formats to improve interoperability and simplify analysis. It also streamlines the import of geospatial data and corrects common issues found in these data sources upon loading.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/read.abares/}
  \item \url{https://codeberg.org/ropensci/read.abares}
  \item Report bugs at \url{https://codeberg.org/ropensci/read.abares/issues}
}

}
\author{
\strong{Maintainer}: Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Other contributors:
\itemize{
  \item Curtin University (\href{https://ror.org/02n415q13}{ROR}) (http://www.curtin.edu.au/) [funder, copyright holder]
  \item Jacob Wujciak-Jens \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID}) (Assisted with troubleshooting formatting in documentation to display '<' and '>' properly) [contributor]
  \item Nicholas Potter \email{econpotter@gmail.com} (\href{https://orcid.org/0000-0002-3410-3732}{ORCID}) (rOpenSci Review https://github.com/ropensci/software-review/issues/667) [reviewer]
  \item María Paula Caldas \email{mpaulacaldas@gmail.com} (\href{https://orcid.org/0000-0002-1938-6471}{ORCID}) (rOpenSci Review https://github.com/ropensci/software-review/issues/667) [reviewer]
}

}
\keyword{internal}
