% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aagis_regions.R
\name{read_aagis_regions}
\alias{read_aagis_regions}
\title{Read 'Australian Agricultural and Grazing Industries Survey' (AAGIS) region mapping files}
\source{
\url{https://www.agriculture.gov.au/sites/default/files/documents/aagis_asgs16v1_g5a.shp_.zip}
}
\usage{
read_aagis_regions(cache = TRUE)
}
\arguments{
\item{cache}{Cache the \acronym{AAGIS} regions' geospatial file after
downloading using \code{tools::R_user_dir("read.abares", "cache")} to identify the
proper directory for storing user data in a cache for this package. Defaults
to \code{TRUE}, caching the files locally as a GeoPackage. If \code{FALSE}, this
function uses \code{tempdir()} and the files are deleted upon closing of the
active \R session.}
}
\value{
An \CRANpkg{sf} object of the \acronym{AAGIS} regions and a
cached geopackage file if requested.
}
\description{
Download, cache and import the Australian Agricultural and Grazing
Industries Survey (\acronym{AAGIS}) regions geospatial shapefile.
}
\details{
@note Upon import a few operations are carried out,
\itemize{
\item the geometries are automatically corrected to fix invalid geometries that
are present in the original shapefile,
\item a new column, "State" is added to be used for mapping state estimates with
data provided by \code{\link[=read_historical_state_estimates]{read_historical_state_estimates()}}, and
\item column names are set to start with an upper-case letter and "name" is set
to "AAGIS_region" to align with the \code{\link[data.table:data.table]{data.table::data.table()}} provided by
\code{\link[=read_historical_regional_estimates]{read_historical_regional_estimates()}} to allow for easier merging of data.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aagis <- read_aagis_regions()

plot(aagis)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/research-topics/surveys/farm-definitions-methods#regions}
}
\concept{AAGIS}
