% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aagis_regions.R
\name{read_aagis_regions}
\alias{read_aagis_regions}
\title{Read 'Australian Agricultural and Grazing Industries Survey' (AAGIS) region mapping files}
\source{
\url{https://www.agriculture.gov.au/sites/default/files/documents/aagis_asgs16v1_g5a.shp_.zip}
}
\usage{
read_aagis_regions(cache = FALSE)
}
\arguments{
\item{cache}{Cache the files after download? Defaults to \code{FALSE} with files
being downloaded to \code{tempdir()} being available throughout the active \R
session and cleaned up on exit. If set to \code{TRUE}, files will be cached
locally for use between sessions. See \dQuote{Caching} section for more.}
}
\value{
An \CRANpkg{sf} object of the \acronym{AAGIS} regions and a
cached geopackage file if requested.
}
\description{
Download, cache and import the Australian Agricultural and Grazing
Industries Survey (\acronym{AAGIS}) regions geospatial shapefile.
}
\details{
@note Upon import a few operations are carried out,
\itemize{
\item the geometries are automatically corrected to fix invalid geometries that
are present in the original shapefile,
\item column names are set to start with an upper-case letter,
\item the original column named, "name", is set to "AAGIS_region" to align with
column names that the \code{\link[data.table:data.table]{data.table::data.table()}} provided by
\code{\link[=read_historical_regional_estimates]{read_historical_regional_estimates()}} to allow for easier merging of data
for mapping, and,
\item a new column, "State" is added to be used for mapping state estimates with
data for mapping state historical estimate values found in the
\code{\link[data.table:data.table]{data.table::data.table()}} from \code{\link[=read_historical_state_estimates]{read_historical_state_estimates()}}.
}
}
\section{Caching}{


When caching is enabled, the directory defined by \link[tools:userdir]{tools::R_user_dir} will
be used by default unless overidden using
\code{read.abares_options(cache_location = "your_desired_file_path")} to change
the file path used for caching. You may also set this globally to \code{TRUE} or
\code{FALSE} by using \code{read.abares_options(read.abares.cache = TRUE)} or
\code{read.abares_options(read.abares.cache = FALSE)}, respectively. Using the
argument in this function will override any options that are set globally.
See \link{read.abares-options} for more.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
aagis <- read_aagis_regions()

plot(aagis)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/research-topics/surveys/farm-definitions-methods#regions}
}
\concept{AAGIS}
