% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abares_trade.R
\name{read_abares_trade}
\alias{read_abares_trade}
\title{Read Data from the ABARES Trade Dashboard}
\source{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0}
}
\usage{
read_abares_trade()
}
\value{
A \CRANpkg{data.table} object of the \acronym{ABARES} trade data.
}
\description{
Fetches and imports \acronym{ABARES} trade data. As the data file is large,
~1.4GB uncompressed \acronym{CSV} file.
}
\note{
Columns are renamed for consistency with other \acronym{ABARES} products
serviced in this package using a snake_case format and ordered
consistently.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
trade <- read_abares_trade()

trade
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/research-topics/trade/dashboard}
}
\seealso{
Other Trade: 
\code{\link{read_abares_trade_regions}()}
}
\concept{Trade}
