% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abares_trade.R
\name{read_abares_trade}
\alias{read_abares_trade}
\title{Read data from the ABARES Trade Dashboard}
\source{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0}
}
\usage{
read_abares_trade(cache = TRUE)
}
\arguments{
\item{cache}{\code{Boolean} Cache the \acronym{ABARES} trade data locally after
download to save download time in the future. Uses \code{tools::R_user_dir()} to
identify the proper directory for storing user data in a cache for this
package. Defaults to \code{TRUE}, caching the file as a gzipped CSV file. If
\code{FALSE}, this function uses \code{tempdir()} and the files are deleted upon
closing of the active \R session.}
}
\value{
A \CRANpkg{data.table} object of the \acronym{ABARES} trade data.
}
\description{
Fetches and imports \acronym{ABARES} trade data.
}
\note{
Columns are renamed for consistency with other \acronym{ABARES} products
serviced in this package using a snake_case format and ordered
consistently.

The cached file is not the same as the raw file that is available for
download. It will follow the renaming scheme and filling values that this
function will perform on the raw data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
trade <- read_abares_trade()

trade
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/research-topics/trade/dashboard}
}
\seealso{
Other Trade: 
\code{\link{read_abares_trade_regions}()}
}
\concept{Trade}
