% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abares_trade.R
\name{read_abares_trade}
\alias{read_abares_trade}
\title{Read data from the ABARES Trade Dashboard}
\source{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0}
}
\usage{
read_abares_trade(cache = getOption("read.abares.cache"))
}
\arguments{
\item{cache}{Cache the files after download? Defaults to \code{FALSE} with files
being downloaded to \code{tempdir()} being available throughout the active \R
session and cleaned up on exit. If set to \code{TRUE}, files will be cached
locally for use between sessions. See \dQuote{Caching} section for more.}
}
\value{
A \CRANpkg{data.table} object of the \acronym{ABARES} trade data.
}
\description{
Fetches and imports \acronym{ABARES} trade data. As the data file is large,
~1.4GB uncompressed CSV file, caching is offered to save repeated
downloading.
}
\note{
Columns are renamed for consistency with other \acronym{ABARES} products
serviced in this package using a snake_case format and ordered
consistently.

The cached file is not the same as the raw file that is available for
download. It will follow the renaming scheme and filling values that this
function will perform on the raw data.
}
\section{Caching}{


When caching is enabled, the directory defined by \code{\link[tools:userdir]{tools::R_user_dir()}} will
be used by default unless overidden using
\code{read.abares_options(cache_location = "your_desired_file_path")} to change
the file path used for caching. You may also set this globally to \code{TRUE} or
\code{FALSE} by using \code{read.abares_options(read.abares.cache = TRUE)} or
\code{read.abares_options("read.abares.cache" = TRUE)}, respectively. Using the
argument in this function will override any options that are set globally.
See \link{read.abares-options} for more.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
trade <- read_abares_trade()

trade
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/research-topics/trade/dashboard}
}
\seealso{
Other Trade: 
\code{\link{read_abares_trade_regions}()}
}
\concept{Trade}
