% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abares_trade.R
\name{read_abares_trade}
\alias{read_abares_trade}
\title{Read Data from the ABARES Trade Dashboard}
\source{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0}
}
\usage{
read_abares_trade(x = NULL, code_description = FALSE)
}
\arguments{
\item{x}{A file path providing the file with the data to be imported. The
file is assumed to be unarchived (\emph{i.e.}, still zipped). This function does
not provide any checking whether this function is the proper function for
the provided file.  Defaults to \code{NULL}, assuming that the file will be
downloaded in the active \R session.}

\item{code_description}{Boolean. Include the trade code description, this
results in a larger object with long text descriptions for each trade code,
and the unit of quantity measure, \emph{e.g.},  "KG" (kilograms), "NO"
(number/count), "L" (litres).}
}
\value{
A \CRANpkg{data.table} object of the \acronym{ABARES} trade data
with the "Trade_code" field as the key for fast subsetting.
}
\description{
Fetches and imports \acronym{ABARES} trade data. As the data, \code{x}, is large,
~1.5GB uncompressed \acronym{CSV}, downloads may be slow. The final object
is sorted by "Fiscal_year" and "Month".
}
\note{
Columns are renamed for consistency with other \acronym{ABARES} products
serviced in this package using a snake_case format and ordered
consistently. "State" reflects the state of origin of the exported
goods, not the state in which the port of departure is located; it is
therefore expected that "State" and "Australian_port" will not always
correspond geographically.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
trade <- read_abares_trade()

trade

trade_codes <- read_abares_trade(code_description = TRUE)

trade_codes
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/research-topics/trade/dashboard}
}
\seealso{
Other Trade:
\code{\link[=read_abares_trade_regions]{read_abares_trade_regions()}}
}
\concept{Trade}
