% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abares_trade_regions.R
\name{read_abares_trade_regions}
\alias{read_abares_trade_regions}
\title{Read ABARES 'Trade Data Regions' from the ABARES Trade Dashboard}
\source{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/2}
}
\usage{
read_abares_trade_regions()
}
\value{
A \CRANpkg{data.table} object of the \acronym{ABARES} trade data
regions.
}
\description{
Fetches and imports \acronym{ABARES} trade regions data.
}
\note{
Columns are renamed for consistency with other \acronym{ABARES} products
serviced in this package using a snake_case format and ordered consistently.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
trade_regions <- read_abares_trade_regions()

trade_regions
\dontshow{\}) # examplesIf}
}
\references{
\url{https://daff.ent.sirsidynix.net.au/client/en_AU/search/asset/1033841/0}
}
\seealso{
Other Trade: 
\code{\link{read_abares_trade}()}
}
\concept{Trade}
