% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agdf_terra.R
\name{read_agdf_terra}
\alias{read_agdf_terra}
\title{Read AGDF NCDF Files}
\usage{
read_agdf_terra(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \CRANpkg{terra} \link[terra:rast]{terra::rast} object of the Australian Gridded Farm
Data
}
\description{
Read AGDF NCDF Files
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agdf_terra()
\dontshow{\}) # examplesIf}
}
\seealso{
Other read_agdf: 
\code{\link{read_agdf_dt}()},
\code{\link{read_agdf_stars}()},
\code{\link{read_agdf_tidync}()}
}
\concept{read_agdf}
