% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agdf_tidync.R
\name{read_agdf_tidync}
\alias{read_agdf_tidync}
\title{Read AGDF NCDF Files with tidync}
\usage{
read_agdf_tidync(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \code{list} of \CRANpkg{tidync} \link[tidync:tidync]{tidync::tidync} objects of the
Australian Gridded Farm Data
}
\description{
Read AGDF NCDF Files with tidync
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agdf_tidync()
\dontshow{\}) # examplesIf}
}
\seealso{
Other read_agdf: 
\code{\link{read_agdf_dt}()},
\code{\link{read_agdf_stars}()},
\code{\link{read_agdf_terra}()}
}
\concept{read_agdf}
