% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_table.R, R/read_tibble.R
\name{read_agfd_dt}
\alias{read_agfd_dt}
\title{Read agfd NCDF Files as a data.table}
\usage{
read_agfd_dt(files)

read_agfd_dt(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \code{list} of \CRANpkg{tidync} \link[tidync:tidync]{tidync::tidync} objects of the
Australian Gridded Farm Data

a \code{list} of \CRANpkg{tidync} \link[tidync:tidync]{tidync::tidync} objects of the
Australian Gridded Farm Data
}
\description{
Read agfd NCDF Files as a data.table

Read agfd NCDF Files as a tibble
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agfd_dt()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agfd_dt()
\dontshow{\}) # examplesIf}
}
\seealso{
Other read_agfd: 
\code{\link{read_agfd_stars}()},
\code{\link{read_agfd_terra}()},
\code{\link{read_agfd_tidync}()}

Other read_agfd: 
\code{\link{read_agfd_stars}()},
\code{\link{read_agfd_terra}()},
\code{\link{read_agfd_tidync}()}
}
\concept{read_agfd}
