% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agdf_stars.R
\name{read_agfd_stars}
\alias{read_agfd_stars}
\title{Read AGFD NCDF Files}
\usage{
read_agfd_stars(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \code{list} of \CRANpkg{stars} objects of the Australian Gridded Farm
Data with filenames as the list's object's names
}
\description{
Read Australian Gridded Farm Data, (\acronym{AGFD}) as a \link{stars} object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agfd_stars()
\dontshow{\}) # examplesIf}
}
\seealso{
Other read_agfd: 
\code{\link{read_agfd_dt}()},
\code{\link{read_agfd_terra}()},
\code{\link{read_agfd_tidync}()}
}
\concept{read_agfd}
