% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agdf_terra.R
\name{read_agfd_terra}
\alias{read_agfd_terra}
\title{Read agfd NCDF Files}
\usage{
read_agfd_terra(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \CRANpkg{terra} \link[terra:rast]{terra::rast} object of the Australian Gridded Farm
Data
}
\description{
Read agfd NCDF Files
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agfd_terra()
\dontshow{\}) # examplesIf}
}
\seealso{
Other read_agfd: 
\code{\link{read_agfd_dt}()},
\code{\link{read_agfd_stars}()},
\code{\link{read_agfd_tidync}()}
}
\concept{read_agfd}
