% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agfd_tidync.R
\name{read_agfd_tidync}
\alias{read_agfd_tidync}
\title{Read agfd NCDF Files with tidync}
\usage{
read_agfd_tidync(files)
}
\arguments{
\item{files}{A list of NetCDF files to import}
}
\value{
a \code{list} object of \CRANpkg{tidync} \link[tidync:tidync]{tidync::tidync} objects of the
Australian Gridded Farm Data with the file names as the  list's objects'
names.
}
\description{
Read agfd NCDF Files with tidync
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_agfd(cache = TRUE) |>
  read_agfd_tidync()
\dontshow{\}) # examplesIf}
}
\seealso{
Other read_agfd: 
\code{\link{read_agfd_dt}()},
\code{\link{read_agfd_stars}()},
\code{\link{read_agfd_terra}()}
}
\concept{read_agfd}
