% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clum_stars.R
\name{read_clum_stars}
\alias{read_clum_stars}
\title{Read catchment scale Land Use of Australia GeoTIFFs using stars}
\source{
\url{https://doi.org/10.25814/2w2p-ph98}
}
\usage{
read_clum_stars(
  data_set = "clum_50m_2023_v2",
  cache = getOption("read.abares.cache"),
  user_agent = getOption("read.abares.user_agent"),
  max_tries = getOption("read.abares.max_tries"),
  timout = getOption("read.abares.max_tries"),
  files = NULL
)
}
\arguments{
\item{data_set}{A string value indicating the data desired for download.
One of:
\describe{
\item{clum_50m_2023_v2}{ (default)},
\item{scale_date_update}{}
}.}

\item{cache}{Boolean cache the files after download? Defaults to \code{FALSE}
with files being downloaded to \code{tempdir()} being available throughout the
active \R session and cleaned up on exit.  If set to \code{TRUE}, files will be
cached locally for use between sessions.  See \dQuote{Caching} section for
more.}

\item{user_agent}{Character string providing a custom user agent to pass
along with the download request for the server logs.}

\item{max_tries}{Integer providing the number of times that a file download
should be retried upon failure.  Defaults to 3.}

\item{files}{An optional \code{\link[=list]{list()}} object of file paths that provide the
location of local files to import.  If not provided, defaults to downloading
the data from \acronym{ABARES}.}
}
\value{
a \CRANpkg{stars} object that may be one or many layers depending
upon the requested data set.
}
\description{
Download and import catchment scale Land Use of Australia GeoTIFFs using
\CRANpkg{stars} object.  Data can be cached on request.
}
\details{
From the
\href{https://www.agriculture.gov.au/sites/default/files/documents/CLUM_DescriptiveMetadata_December2023_v2.pdf}{ABARES documentation}
\dQuote{The Catchment Scale Land Use of Australia – Update December 2023
version 2 dataset is the national compilation of catchment scale land use
data available for Australia (CLUM), as at December 2023. It replaces the
Catchment Scale Land Use of Australia – Update December 2020. It is a
seamless raster dataset that combines land use data for all state and
territory jurisdictions, compiled at a resolution of 50 metres by 50 metres.
The CLUM data shows a single dominant land use for a given area, based on the
primary management objective of the land manager (as identified by state and
territory agencies). Land use is classified according to the Australian Land
Use and Management Classification version 8. It has been compiled from vector
land use datasets collected as part of state and territory mapping programs
and other authoritative sources, through the Australian Collaborative Land
Use and Management Program. Catchment scale land use data was produced by
combining land tenure and other types of land use information including,
fine-scale satellite data, ancillary datasets, and information collected in
the field. The date of mapping (2008 to 2023) and scale of mapping (1:5,000
to 1:250,000) vary, reflecting the source data, capture date and scale.
Date and scale of mapping are provided in supporting datasets.}
-- \acronym{ABARES}, 2024-06-27
}
\note{
The raster will load with the default category for each data set, but you can
specify a different category to use after loading.
\code{active_cat} argument. To see which categories are available, please refer
to the metadata for these data.  The PDF can be accessed in your default web
browser by using \code{\link[=view_clum_metadata_pdf]{view_clum_metadata_pdf()}}.
}
\section{Caching}{


When caching is enabled, the directory defined by \code{\link[tools:userdir]{tools::R_user_dir()}} will
be used by default unless overidden using
\code{read.abares_options(cache_location = "your_desired_file_path")} to change
the file path used for caching. You may also set this globally to \code{TRUE} or
\code{FALSE} by using \code{read.abares_options(read.abares.cache = TRUE)} or
\code{read.abares_options("read.abares.cache" = TRUE)}, respectively. Using the
argument in this function will override any options that are set globally.
See \link{read.abares-options} for more.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

cs <- read_clum_stars(data_set = "clum_50m_2023_v2")

cs

plot(cs)
\dontshow{\}) # examplesIf}
}
\references{
ABARES 2024, Catchment Scale Land Use of Australia – Update December 2023
version 2, Australian Bureau of Agricultural and Resource Economics and
Sciences, Canberra, June, CC BY 4.0, DOI: \doi{10.25814/2w2p-ph98}
}
\seealso{
Other clum: 
\code{\link{read_clum_commodities}()},
\code{\link{read_clum_terra}()}
}
\concept{clum}
