% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_estimates_by_performance_category.R
\name{read_estimates_by_performance_category}
\alias{read_estimates_by_performance_category}
\alias{read_est_by_perf_cat}
\title{Read Estimates by Size From ABARES}
\source{
\url{https://www.agriculture.gov.au/sites/default/files/documents/fdp-BySize-ByPerformance.csv}
}
\usage{
read_estimates_by_performance_category()

read_est_by_perf_cat()
}
\value{
A \link[data.table:data.table]{data.table::data.table} object
}
\description{
Read Estimates by Size From ABARES
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

 read_estimates_by_performance_category()

 # or shorter
 read_est_by_perf_cat()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.agriculture.gov.au/abares/data/farm-data-portal#data-download}
}
\seealso{
Other Estimates: 
\code{\link{read_estimates_by_size}()},
\code{\link{read_historical_national_estimates}()},
\code{\link{read_historical_regional_estimates}()},
\code{\link{read_historical_state_estimates}()}
}
\concept{Estimates}
