% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nlum_stars.R
\name{read_nlum_stars}
\alias{read_nlum_stars}
\title{Read national scale Land Use of Australia GeoTIFFs using stars}
\source{
\url{https://doi.org/10.25814/w175-xh85}
}
\usage{
read_nlum_stars(data_set, active_cat, cache = TRUE)
}
\arguments{
\item{data_set}{A string value indicating the GeoTIFF desired for download.
One of:
\describe{
\item{Y201011}{Land use of Australia 2010–11}
\item{Y201516}{Land use of Australia 2015–16}
\item{Y202021}{Land use of Australia 2020–21}
\item{C201021}{Land use of Australia change}
\item{T201011}{Land use of Australia 2010–11 thematic layers}
\item{T201516}{Land use of Australia 2015–16 thematic layers}
\item{T202021}{Land use of Australia 2020–21 thematic layers}
\item{P201011}{Land use of Australia 2010–11 agricultural commodities probability grids}
\item{P201516}{Land use of Australia 2015–16 agricultural commodities probability grids}
\item{P202021}{Land use of Australia 2020–21 agricultural commodities probability grids}
}}

\item{cache}{Cache the Australian Gridded Farm Data files after download
using \link[tools:userdir]{tools::R_user_dir} to identify the proper directory for storing user
data in a cache for this package. Defaults to \code{TRUE}, caching the files
locally. If \code{FALSE}, this function uses \code{tempdir()} and the files are
deleted upon closing of the active \R session.}
}
\description{
Download and import national scale Land Use of Australia GeoTIFFs using
\CRNApkg{stars} object.  Data are cached on request.
}
\details{
From the
\href{https://www.agriculture.gov.au/abares/aclump/land-use/land-use-of-australia-2010-11-to-2020-21}{ABARES website}:
\dQuote{The \emph{Land use of Australia 2010–11 to 2020–21} data package consists
of seamless continental rasters that present land use at national scale for
2010–11, 2015–16 and 2020–21 and the associated change between each target
period.  Non-agricultural land uses are mapped using 7 thematic layers,
derived from existing datasets provided by state and territory jurisdictions
and external agencies. These 7 layers are: protected areas, topographic
features, land tenure, forest type, catchment scale land use, urban
boundaries, and stock routes. The agricultural land uses are based on the
Australian Bureau of Statistics’ 2010–11, 2015–16 and 2020–21 agricultural
census data; with spatial distributions modelled using Terra Moderate
Resolution Imaging Spectroradiometer (\acronym{MODIS}) satellite imagery and
training data, assisted by spatial constraint layers for cultivation,
horticulture, and irrigation.
Land use is specified according to the Australian Land Use and Management
(\acronym{ALUM}) Classification version 8. The same method is applied to all
target periods using representative national datasets for each period, where
available. All rasters are in GeoTIFF format with geographic coordinates in
Geocentric Datum of Australian 1994 (GDA94) and a 0.002197 degree
(~250 metre) cell size.
The \emph{Land use of Australia 2010–11 to 2020–21} data package is a product of
the Australian Collaborative Land Use and Management Program. This data
package replaces the Land use of Australia 2010–11 to 2015–16 data package,
with updates to these time periods.}
-- \acronym{ABARES}, 2024-11-28

The raster will load with the default category for each data set, but you can
specify a different category to use when reading the file in with the
\code{active_cat} argument. To see which categories are available, please refer
to the metadata for these data.  The PDF can be accessed in your default web
browser by using \code{\link[=view_nlum_metadata_pdf]{view_nlum_metadata_pdf()}}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

read_nlum_stars()

nlum_stars

plot(nlum_stars)
\dontshow{\}) # examplesIf}
}
\references{
ABARES 2024, Land use of Australia 2010–11 to 2020–21, Australian Bureau of
Agricultural and Resource Economics and Sciences, Canberra, November, CC BY
4.0. \doi{10.25814/w175-xh85}
}
\seealso{
Other nlum: 
\code{\link{read_nlum_terra}()},
\code{\link{view_nlum_metadata_pdf}()}
}
\concept{nlum}
