% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_topsoil_thickness_stars.R
\name{read_topsoil_thickness_stars}
\alias{read_topsoil_thickness_stars}
\title{Read a 'Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1' file with stars}
\source{
\url{https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip}
}
\usage{
read_topsoil_thickness_stars(files = NULL)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[terra:rast]{terra::rast()}}, for \emph{e.g.},
\code{activeCat} if you wished to set the active category when loading any of the
available GeoTIFF files that are encoded with a raster attribute table.}
}
\value{
A \link{stars} object of the 'Soil Thickness for Australian Areas of
Intensive Agriculture of Layer 1'.
}
\description{
Read Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
data as a \CRANpkg{stars} object.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
st_stars <- read_topsoil_thickness_stars()

plot(st_stars)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
}
\seealso{
Other topsoil thickness: 
\code{\link{.get_topsoil_thickness}()},
\code{\link{print_topsoil_thickness_metadata}()},
\code{\link{read_topsoil_thickness_terra}()}
}
\concept{topsoil thickness}
