% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_topsoil_thickness_terra.R
\name{read_topsoil_thickness_terra}
\alias{read_topsoil_thickness_terra}
\title{Read a Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1 file with terra}
\source{
\url{https://anrdl-integration-web-catalog-saxfirxkxt.s3-ap-southeast-2.amazonaws.com/warehouse/staiar9cl__059/staiar9cl__05911a01eg_geo___.zip}
}
\usage{
read_topsoil_thickness_terra(
  cache = getOption("read.abares.cache"),
  user_agent = getOption("read.abares.user_agent"),
  max_tries = getOption("read.abares.max_tries"),
  timout = getOption("read.abares.max_tries"),
  files = NULL
)
}
\arguments{
\item{cache}{Boolean cache the files after download? Defaults to \code{FALSE}
with files being downloaded to \code{tempdir()} being available throughout the
active \R session and cleaned up on exit.  If set to \code{TRUE}, files will be
cached locally for use between sessions.  See \dQuote{Caching} section for
more.}

\item{user_agent}{Character string providing a custom user agent to pass
along with the download request for the server logs.}

\item{max_tries}{Integer providing the number of times that a file download
should be retried upon failure.  Defaults to 3.}

\item{files}{An optional \code{\link[=list]{list()}} object of file paths that provide the
location of local files to import.  If not provided, defaults to downloading
the data from \acronym{ABARES}.}

\item{...}{Additional arguments passed to \code{\link[terra:rast]{terra::rast()}}, for \emph{e.g.},
\code{activeCat} if you wished to set the active category when loading any of the
available GeoTIFF files that are encoded with a raster attribute table.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object of the 'Soil Thickness for Australian Areas
of Intensive Agriculture of Layer 1'.
}
\description{
Read Soil Thickness for Australian Areas of Intensive Agriculture of Layer 1
data as a \code{\link[terra:rast]{terra::rast()}} object.
}
\section{Caching}{


When caching is enabled, the directory defined by \code{\link[tools:userdir]{tools::R_user_dir()}} will
be used by default unless overidden using
\code{read.abares_options(cache_location = "your_desired_file_path")} to change
the file path used for caching. You may also set this globally to \code{TRUE} or
\code{FALSE} by using \code{read.abares_options(read.abares.cache = TRUE)} or
\code{read.abares_options("read.abares.cache" = TRUE)}, respectively. Using the
argument in this function will override any options that are set globally.
See \link{read.abares-options} for more.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
st_terra <- read_topsoil_thickness_terra()

# terra::plot() is reexported for convenience
plot(st_terra)
\dontshow{\}) # examplesIf}
}
\references{
\url{https://data.agriculture.gov.au/geonetwork/srv/eng/catalog.search#/metadata/faa9f157-8e17-4b23-b6a7-37eb7920ead6}
}
\seealso{
Other topsoil thickness: 
\code{\link{get_topsoil_thickness}()},
\code{\link{print_topsoil_thickness_metadata}()},
\code{\link{read_topsoil_thickness_stars}()}
}
\concept{topsoil thickness}
