% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_nlum_metadata_pdf.R
\name{view_nlum_metadata_pdf}
\alias{view_nlum_metadata_pdf}
\title{Displays the PDF metadata for the National Land Use raster files in a native viewer}
\usage{
view_nlum_metadata_pdf()
}
\value{
Called for its side-effects, opens the system's native PDF viewer to
display the requested metadata file.
}
\description{
Each National Land Use raster file comes with a PDF of metadata. This
function will open and display that file using the native PDF viewer for any
system.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
view_nlum_metadata_pdf()
\dontshow{\}) # examplesIf}
}
